Main

Description

Tarefa 1 da 1ª Fase do projeto de LI1.

Funções de teste

correTestes :: IO ()

Corre múltiplos testes para a tarefa 1.

correTeste :: ([String] -> [String]) -> String -> IO ()

Corre um teste para uma tarefa.

Funções da Tarefa 1

inStr :: String -> [String]

Esta função recebe uma string e transforma-a numa lista de strings, dividindo-a quando encontrar um '\n'.

outStr :: [String] -> String

Esta função recebe uma lista de strings e transforma-a numa única string, adicionando '\n' entre as strings originais.

main :: IO ()

Função geral da tarefa 1.

tarefa1 :: [String] -> [String]

Esta função recebe um tabuleiro e verifica se este está no formato correto, devolvendo ["OK"], em caso afirmativo, ou a linha onde está o primeiro erro, em caso contrário. Seguem-se dois exemplos:

 >>> tarefa1 ["#####","# ###","#####","1 1"]
 ["OK"]

 >>> tarefa1 ["#####","# ..#","#   #","#####","0 0","2 2"]
 ["5"]

tarefa1Erro :: [String] -> Int

Esta função recebe um tabuleiro e devolve a linha com erro, se houver e caso contrário, devolve -1.

 >>> tarefa1Erro ["#####","# ###","#####","1 1"]
 -1

 >>> tarefa1Erro ["#####","# ..#","#   #","#####","0 0","2 2"]
 5

Funções auxiliares

juntaErros :: Int -> Int -> Int

Função que devolve a linha menor dos erros encontrados. Se não foram encontrados erros devolve -1.

 >>> juntaErros (-1) 3
 3

 >>> juntaErros (-1) (-1)
 -1

 >>> juntaErros 1 4
 1  

processaCoordenadas :: Int -> [String] -> Int

Verifica que se as coordenadas estão no formato pedido.

 >>> processaCoordenadas 7 ["1 1","3 4"]
-1

 >>> processaCoordenadas 7 ["1 a"]
7

 >>> processaCoordenadas 7 ["1 1","3  "]
8

processa :: [String] -> [String]

Função que apaga eventuais listas vazias no final do tabuleiro.

>>> processa ["#####","#   #","##. #","#####","     "]
["#####","#   #","##. #","#####"]

parteMapa :: [String] -> ([String], [String])

Função que divide a parte das coordenadas da parte do tabuleiro.

 >>> parteMapa ["#####","# ###","#####","1 1"]
 (["#####","# ###","#####"],["1 1"])

lerCoordenadas :: [String] -> [(Int, Int)]

Caso o processaCoordenadas verifique que as coordenadas estão formato pedido, esta função transforma as coordenadas no formato (x,y).

 >>> lerCoordenadas ["1 1","3 4"]
[(1,1),(3,4)]

Funções que verificam o tabuleiro

validaTab :: [String] -> Int

Verifica se o tabuleiro está bem construído.

 >>> validaTab ["#####","# ###","#####"]
 -1

 >>> validaTab ["# ###","# ###","#####"]
 1

validaTab1 :: Int -> [String] -> Int

Verifica se o ficheiro é apenas composto pelos caracteres '#', ' ' ou '.' e se cada linha acaba e começa em '#'.

 >>> validaTab1 1 ["#####","# ###","#####"]
 -1

 >>> validaTab1 1 ["#####","# ###","##2##"]
 3

 >>> validaTab1 1 ["#####","  ###","#####"]
 2

validaTab2 :: [String] -> Int

Verifica se a primeira e a última linha são constituídas apenas por '#'.

 >>> validaTab2 ["#####","# ###","#####"]
 -1

 >>> validaTab2 ["#####","# ###","## ##"]
 3

validaTab3 :: Int -> [String] -> Int

Verifica se as linhas têm todas o mesmo comprimento.

 >>> validaTab3 2 ["#####","# ###","#####"]
 -1

 >>> validaTab3 2 ["#####","# ##","## ##"]
 2

Funções que verificam as coordenadas

validaCoords :: [(Int, Int)] -> [String] -> Int

Verifica se as coordenadas estão em locais possíveis do tabuleiro:

não estão fora do tabuleiro;

não estão em cima de paredes, etc...

 >>> validaCoords  [(1,1)] ["#####","# ###","#####"]
 -1

 >>> validaCoords [(3,4)] ["# ###","# ###","#####"] 
 4

validaCoords1 :: Int -> [(Int, Int)] -> [String] -> Int

Verifica se todos os pares de coordenadas se encontram dentro do tabuleiro.

 >>> validaCoords1 7 [(1,1)] ["#####","# ###","#####"]
 -1

 >>> validaCoords1 7 [(3,4)] ["#####","# ###","#####"]
 7

validaCoords2 :: Int -> [(Int, Int)] -> [String] -> Int

Verifica se todas as coordenadas estão num '.' ou num ' '.

 >>> validaCoords2 7 [(1,1)] ["#####","# ###","#####"]
 -1

 >>> validaCoords2 7 [(2,2)] ["#####","# ###","#####"]
 7

validaCoords3 :: [(Int, Int)] -> [String] -> Int

Verifica se o tabuleiro tem tantos locais de arrumações ('.') quanto o número de caixas.

 >>> validaCoords3 [(1,1),(1,2)] ["#####","# ###","#.###","#####"]
 -1

 >>> validaCoords3 [(1,1)] ["#####","# ###","#.###","#####"]
 6

validaCoords4 :: Int -> [(Int, Int)] -> Int

Verifica se a parte das coordenadas não tem coordenadas repetidas.

 >>> validaCoords4 7 [(2,3),(3,7),(4,5)]
 -1

 >>> validaCoords4 7 [(2,3),(3,7),(2,3)]
 9