Sokoban

Description

Módulo inicial da resolução correspondente à Tarefa F da segunda fase do projeto de Laboratórios de Informática I.

Documentation

main :: IO ()

Função inicial do programa.

data State

Armazena todo o estado do programa.

Constructors

State

Fields

stateShared :: Shared

stateStageMgr :: StageMgr

createState

Arguments

:: VectorI
Resolução inicial da janela.

-> IO State

Cria o estado inicial do programa.

handleEvent

Arguments

:: Event
Evento.

-> State

-> IO State

Processa um evento.

update

Arguments

:: Float
Tempo desde a última atualização, em segundos.

-> State

-> IO State

Atualiza o estado do programa.

draw :: State -> IO Picture

Transforma o estado do programa num valor do tipo Picture para ser desenhado no ecrã.