Sokoban.Assets

Description

Este módulo é destinado ao carregamento e acesso a recursos externos utilizados pelo jogo, e.g. imagens e tipos de letra.

Constantes

Constantes utilizadas pelo módulo.

uiDir :: FilePath

Caminho relativo da diretoria onde se encontram os bitmaps relativos à UI.

fontDir :: FilePath

Caminho relativo da diretoria onde se encontram os ficheiros relativos ao tipo de letra utilizado pelo jogo.

playerDir :: FilePath

Caminho relativo da diretoria onde se encontram os bitmaps associados com cada personagem utilizados pelo jogo.

themeDir :: FilePath

Caminho relativo da diretoria onde se encontram os bitmaps associados com cada tema (aspeto gráfico do tabuleiro e caixas) utilizados pelo jogo.

Assets

Tipos e funções relacionados com assets em geral.

Define-se um asset como sendo um recurso externo, e.g. uma imagem.

data Assets

Tipo abstrato que armazena todos os recursos externos utilizados pelo jogo.

Constructors

Assets

Fields

assetsUI :: UIAssets

Assets relacionados com a UI.

assetsPlayers :: PlayerAssets

Assets relacionados com as personagens.

assetsThemes :: [ThemeAssets]

Assets relacionados com as tiles e caixas.

load :: IO Assets

Carrega todos os assets disponíveis, a partir das diretorias predefinidas.

Interface do Utilizador (UI)

Tipos e funções relacionados com assets utilizados pela UI.

data UIAssets

Armazena os bitmaps e o tipo de letra utilizados pela UI.

Constructors

UIAssets

Fields

uiAssetsBitmaps :: BitmapMap

Bitmaps utilizados pela UI.

uiAssetsFont :: Font

Tipo de letra utilizado pela UI.

uiBitmap

Arguments

:: String
Nome do bitmap.

-> Assets
Assets.

-> Picture

Obtém um bitmap da UI.

uiFont :: Assets -> Font

Obtém o tipo de letra da UI.

Carregamento

Funções relacionadas com o carregamento de assets da UI.

loadUI :: IO UIAssets

Carrega os bitmaps e o tipo de letra utilizados pela UI, a partir das diretorias predefinidas.

Personagens

Tipos e funções relacionados com assets utilizados pelas personagens.

type PlayerAssets = Map String BitmapMap

Armazena os bitmaps utilizados pelas personagens.

playerBitmap

Arguments

:: String
Nome da personagem.

-> String
Nome do bitmap.

-> Assets
Assets.

-> Picture

Obtém um bitmap de uma personagem.

Carregamento

Funções relacionadas com o carregamento de assets das personagens.

loadPlayers :: IO PlayerAssets

Carrega os bitmaps utilizados pelas personagens, a partir das diretorias predefinidas.

loadPlayer

Arguments

:: FilePath
Caminho da diretoria. Deve terminar com '/'.

-> IO (String, BitmapMap)
(Nome da personagem, Bitmaps da personagem)

Carrega os bitmaps utilizados por uma personagem, a partir da diretoria especificada.

Tiles e Caixas

Tipos e funções relacionados com assets utilizados pelas tiles e caixas.

Uma tile é uma das células que constituem o tabuleiro de jogo e a respetiva imagem.

newtype Theme

Tipo abstrato que dentifica um tema.

Os temas afetam o aspeto das tiles e das caixas utilizados para desenhar o tabuleiro de jogo. Cada tema é constituido por uma série de sprites que são utilizados para representar as tiles e as caixas.

Constructors

Theme Int

data ThemeAssets

Armazena os sprites utilizados por cada tema.

Constructors

ThemeAssets

Fields

themeAssetsTiles :: SpriteList

Sprites das tiles.

themeAssetsTarget :: SpriteBitmaps

Sprite da tile para locais de arrumação.

themeAssetsBox :: SpriteBitmaps

Sprite para as caixas.

randomTheme :: Assets -> IO Theme

Obtém um identificador de tema aleatório, referente a um dos temas disponíveis no Assets especificado.

tileSprite

Arguments

:: String
Padrão a procurar.

-> SpriteList
Lista de sprites a analisar.

-> SpriteList
Lista de sprites resultante.

Devolve uma lista de todos os sprites de tiles que correspondem ao padrão especificado (cf. tileBitmap).

tileBitmap

Arguments

:: String
Padrão das células.

-> Theme
Identificador do tema a utilizar.

-> Assets
Assets.

-> IO Picture

Obtém um bitmap do sprite da tile correspondente ao padrão especificado.

O sprite que cada célula do tabuleiro terá é dependente da própria célula e das 8 células que a rodeiam. O padrão especificado deve descrever essas 9 células, linha por linha, numa string de 9 carateres. Os carateres permitidos são os seguintes:

'g' - a célula é vazia ou é um local de arrumação;

'w' - a célula é uma parede.

Exemplo: "gwgwwgwwg" corresponde à seguinte disposição de células:

gwg
wwg
wwg

Nota: se a célula central for um local de arrumação, deve ser utilizada a função targetBitmap.

targetBitmap

Arguments

:: Theme
Identificador do tema a utilizar.

-> Assets
Assets.

-> IO Picture

Obtém um bitmap do sprite da tile correspondente a um local de arrumação.

boxBitmap

Arguments

:: Theme
Identificador do tema a utilizar.

-> Assets
Assets.

-> IO Picture

Obtém um bitmap do sprite correspondente às caixas.

Carregamento

Funções relacionadas com o carregamento de assets das tiles e caixas.

loadThemes :: IO [ThemeAssets]

Carrega os assets utilizados pelos temas, a partir das diretorias predefinidas.

loadTheme

Arguments

:: FilePath
Caminho da diretoria. Deve terminar com '/'.

-> IO ThemeAssets

Carrega os assets utilizados por um tema, a partir da diretoria especificada.

checkTiles

Arguments

:: String
Nome do tema.

-> SpriteList
Lista de sprites a verificar.

-> SpriteList

Verifica se uma lista de sprites representa corretamente um conjunto de tiles.

Se a lista for válida, devolve a mesma lista; em caso contrário, causa um erro.

checkTile

Arguments

:: String
Nome do tema.

-> Sprite
Sprite a verificar.

-> Sprite

Verifica se um sprite representa corretamente uma tile.

Se o sprite for válido, devolve o mesmo sprite; em caso contrário, causa um erro.

themeError

Arguments

:: String
Nome do tema.

-> String
Mensagem de erro.

-> a

Causa um erro relacionado com o tema de nome especificado.

Bitmaps

Tipos e funções relacionados com bitmaps em geral.

type BitmapMap = Map String Picture

Mapa que associa um nome a cada Picture que contém um bitmap.

type NamedBitmap = (String, Picture)

Tuplo que associa um nome a uma Picture que contém um bitmap.

Carregamento

Funções relacionadas com o carregamento de bitmaps.

loadBitmapsFromDir

Arguments

:: FilePath
Caminho da diretoria. Deve terminar com '/'.

-> IO BitmapMap

Carrega todos os bitmaps que se encontram na diretoria especificada.

São considerados todos os ficheiros de extensão .bmp. As subdiretorias não são incluídas.

É devolvido um mapa que associa um nome a cada bitmap. O nome de um bitmap corresponde ao nome do ficheiro sem extensão de onde foi carregado.

loadBitmap

Arguments

:: FilePath
Caminho do ficheiro a carregar.

-> IO NamedBitmap

Carrega um bitmap a partir de um ficheiro.

É devolvido um tuplo que associa um nome ao bitmap. O nome do bitmap corresponde ao nome do ficheiro sem extensão.

Sprites

Tipos e funções relacionados com sprites.

Define-se um sprite como sendo uma correspondência entre um nome e um conjunto de bitmaps, cada um com uma probabilidade associada. A soma das probabilidades de todos os bitmaps num sprite deve ser de 100%.

Os sprites destinam-se a oferecer acesso aleatório a um conjunto de bitmaps, dependente da probabilidade de cada bitmap.

type SpriteList = [Sprite]

Uma lista de sprites.

type Sprite = (String, SpriteBitmaps)

Representa um sprite.

Associa um nome a uma lista de bitmaps, cada um com uma probabilidade associada.

A lista de bitmaps corresponde a uma lista de tuplos em que o primeiro elemento representa a probabilidade associada com um bitmap e o segundo elemento representa o bitmap. A probabilidade de cada bitmap deve ter um valor entre 1 e 100, e a soma das probabilidades de todos os bitmaps deve ser 100.

Os elementos da lista de tuplos devem estar ordenados por ordem crescente da probabilidade.

type SpriteBitmaps = [(Int, Picture)]

Lista de bitmaps de um sprite (cf. Sprite).

type SpritePart = (String, (Int, Picture))

Constitui uma parte de um sprite.

Associa um nome com uma probabilidade e um bitmap.

A probabilidade deve ter um valor entre 1 e 100.

bitmapFromSprite :: Sprite -> IO Picture

Devolve aleatoriamente um dos bitmaps de um sprite.

É tomada em consideração a probabilidade associada com cada bitmap.

bitmapFromSpriteBitmaps :: SpriteBitmaps -> IO Picture

Devolve aleatoriamente um dos bitmaps de uma lista de bitmaps de um sprite.

É tomada em consideração a probabilidade associada com cada bitmap.

Carregamento

Funções relacionadas com o carregamento de sprites.

loadSpritesFromDir

Arguments

:: FilePath
Caminho da diretoria. Deve terminar com '/'.

-> IO SpriteList

Constrói uma lista de sprites a partir de todos os bitmaps que se encontram na diretoria especificada.

São considerados todos os ficheiros de extensão .bmp. As subdiretorias não são incluídas.

Os nomes e bitmaps dos sprites resultantes são calculados a partir dos nomes dos ficheiros carregados (cf. loadSpritePart).

loadSprites

Arguments

:: [FilePath]
Caminho dos ficheiros a carregar.

-> IO SpriteList

Constrói uma lista de sprites a partir de todos os bitmaps especificados.

Os nomes e bitmaps dos sprites resultantes são calculados a partir dos nomes dos ficheiros carregados (cf. loadSpritePart).

mergeSpriteParts :: [SpritePart] -> SpriteList

Contrói uma lista de sprites a partir de uma lista de valores do tipo SpritePart.

SpriteParts com o mesmo nome são fundidas para formar um sprite.

checkSprite :: Sprite -> Sprite

Verifica se a probabilidade total de um sprite é de 100%.

Se o sprite for válido, devolve o mesmo sprite; em caso contrário, causa um erro.

loadSpritePart

Arguments

:: FilePath
Caminho do ficheiro a carregar.

-> IO SpritePart

Carrega um valor do tipo SpritePart a partir do ficheiro especificado.

O nome e probabilidade associados com a SpritePart são extraídos do nome do ficheiro, o qual deve seguir um formato específico:

<nome>[.<n>_p<prob>].bmp

<nome> - Nome da SpritePart;

<n> - Sequência de carateres arbitrária; diferencia entre ficheiros de sprites com o mesmo nome e probabilidade;

<prob> - Probabilidade associada com a SpritePart; inteiro entre 1 e 100.

Quando a parte opcional é omitida, a SpritePart toma uma probabilidade de 100%.

Exemplos:

imagem.bmp - SpritePart de nome "imagem" e probabilidade 100;

ola.1_p17.bmp - SpritePart de nome "ola" e probabilidade 17.