Sokoban.Game.Boxes

Description

Este módulo gere o armazenamento e representação visual das caixas que o utilizador deve movimentar para terminar o nível.

Tipos

type Boxes = [Box]

Representa um conjunto de caixas.

data Box

Representa uma caixa.

Constructors

Box

Fields

boxPos :: VectorI

boxAnimPos :: Vector

boxAnimInitPos :: Vector

boxAnimFinalPos :: Vector

boxPic :: Picture

boxSteps :: [Step]

data Step

Armazena informação sobre um movimento de uma caixa.

Constructors

Step

Fields

stepInvDir :: VectorI

Construção

create

Arguments

:: Assets
Assets.

-> Theme
Tema.

-> [VectorI]
Posições iniciais da caixa.

-> IO Boxes

Contrói um valor do tipo Boxes, dadas as posições iniciais de todas as caixas.

createBox

Arguments

:: Assets
Assets.

-> Theme
Tema.

-> VectorI

-> IO Box

Contrói um valor do tipo Box, dada a posição inicial da caixa.

Estado

boxPositions :: Boxes -> [VectorI]

Determina as coordenadas das caixas representadas por um valor do tipo Boxes.

hasBoxInPos

Arguments

:: VectorI
Coordenadas a verificar.

-> Boxes
Caixas.

-> Bool

Determina se existe uma caixa nas coordenadas especificadas num valor do tipo Boxes.

Ações

move

Arguments

:: VectorI
Direção do movimento.

-> [VectorI]
Coordenadas das caixas a movimentar.

-> Boxes

-> Boxes

Move as caixas que se encontram nas posições especificadas por um célula na direção especificada.

fakeMove :: Boxes -> Boxes

Utilizado quando não se quer mover nenhuma caixa, mas se quer registar um movimento nulo, que pode depois ser anulado.

moveBox

Arguments

:: VectorI
Direção do movimento.

-> [VectorI]
Coordenadas das caixas a movimentar.

-> Box

-> Box

Move uma caixa por uma célula na direção especificada, se as coordenadas desta se encontrarem na lista especificada.

undo :: Boxes -> Boxes

Anula o último movimento das caixas.

undoBox :: Box -> Box

Anula o último movimento de uma caixa.

Atualização

update

Arguments

:: Float
Fator, de 0 a 1, que determina o estado da animação atual.

-> Boxes

-> Boxes

Atualiza o estado de um conjunto de caixas.

updateBox

Arguments

:: Float
Fator, de 0 a 1, que determina o estado da animação atual.

-> Box

-> Box

Atualiza o estado de uma caixa.

Conversão para Picture

draw :: Boxes -> Picture

Desenha um conjunto de caixas.

drawBox :: Box -> Picture

Desenha uma caixa.