Sokoban.Game.Buttons

Description

Este módulo gere o painel de botões que permite efetuar ações sobre o estado do jogo, e.g. anular uma jogada ou desistir e passar ao próximo nível.

Documentation

buttonSize :: Vector

Dimensões de cada botão.

buttonWidth :: Float

Largura de cada botão.

buttonHeight :: Float

Altura de cada botão.

buttonSpacing :: Float

Espaço livre entre dois botões.

data Buttons

Tipo abstrato que representa um painel de botões.

Constructors

GameButtonsPlaying

Fields

buttonsPlayingUndo :: Button

Botão que anula uma jogada.

buttonsPlayingRestart :: Button

Botão que recomeça o jogo.

buttonsPlayingGiveUp :: Button

Botão que termina o jogo por desistência.

GameButtonsNext

Fields

buttonsNextNext :: Button

Botão que passa ao próximo nível.

buttonsNextQuit :: Button

Botão que termina a fase do jogo, regressando-se ao menu principal.

GameButtonsFinished

Fields

buttonsFinishedQuit :: Button

Botão que termina a fase do jogo, regressando-se ao menu principal.

data Result

Representa o resultado de premir num dos botões de um painel de botões.

Constructors

None
Nenhum dos botões foi premido.

Undo
Deve-se anular uma jogada.

Restart
O jogo deve recomeçar.

GiveUp
O jogo deve terminar por desistência.

Next
Deve-se passar ao próximo nível.

Quit
Deve-se regressar ao menu principal.

createPlaying

Arguments

:: Assets
Assets.

-> Buttons

Contrói um painel com os botões de anular uma jogada, recomeçar o jogo e terminar o jogo por desistência.

createNext

Arguments

:: Assets
Assets.

-> Buttons

Contrói um painel com os botões de passar ao próximo nível e de regressar ao menu principal.

createFinished

Arguments

:: Assets
Assets.

-> Buttons

Contrói um painel com o botão de regressar ao menu principal.

handleEvent

Arguments

:: Event
Evento.

-> RectMapping
Transformação entre espaço do painel de botões e espaço da janela.

-> Buttons
Painel de botões.

-> (Buttons, Result)

Processa um evento.

É devolvido um tuplo com o novo painel de botões e o resultado do processamento do evento.

drawingRegion :: Rect

Determina a região em que um painel de botões será desenhado.

draw :: Buttons -> Picture

Desenha um painel de botões.