Sokoban.Game.Game

Description

Este módulo gere a fase do jogo, durante a qual o utilizador movimenta um jogador para resolver níveis.

Documentation

data Game

Tipo abstrato que representa a fase do jogo, onde o utilizador pode movimentar o personagem ao longo de vários níveis.

Constructors

Game

Fields

gameLevels :: [String]

Níveis a serem jogados.

gamePlayerAttr :: PlayerAttr

Atributos do personagem a ser utilizado.

gameLevel :: Level

Nível atual.

gameStats :: Stats

Stats.

gameButtons :: Buttons

Botões.

create

Arguments

:: Shared
Estado partilhado.

-> [String]
Níveis a serem jogados.

-> PlayerAttr
Atributos do personagem a utilizar.

-> IO Game

Constrói um valor do tipo Game, dados os níveis a serem jogados e os atributos do personagem a utilizar.

minLevelRegionHeightFactor :: Float

Percentagem mínima da altura da janela que o nível deve ocupar.

dstLevelRegion

Arguments

:: Shared
Estado partilhado.

-> Rect

Calcula a região da janela que o nível ocupa.

dstGameStatsRegion

Arguments

:: Shared
Estado partilhado.

-> Rect

Calcula a região da janela que os stats ocupam.

dstButtonsRegion

Arguments

:: Shared
Estado partilhado.

-> Rect

Calcula a região da janela que os botões ocupam.

advanceLevel

Arguments

:: Assets
Assets.

-> Game

-> IO Game

Avança para o próximo nível.

handleEvent

Arguments

:: Shared
Estado partilhado.

-> Event
Evento.

-> Game

-> IO (Maybe Game)

Processa um evento.

Se o Game resultante já tiver terminado (i.e. está pronto para voltar ao menu), é devolvido o valor Nothing. Em caso contrário, é devolvido o valor Just newGame, onde newGame corresponde ao novo estado do Game.

update

Arguments

:: Shared
Estado partilhado.

-> Float
Tempo desde a última atualização, em segundos.

-> Game

-> IO Game

Atualiza o estado do Game.

draw

Arguments

:: Shared
Estado partilhado.

-> Game

-> Picture

Converte o Game numa Picture.