Sokoban.Game.GameDefs

Description

Este módulo contém constantes e definições relativas à fase do jogo.

Documentation

tileSize :: Num a => (a, a)

Dimensões, em píxeis, de uma tile.

animDuration :: Float

Duração da animação do movimento do jogador e das caixas quando o utilizador movimenta o jogador.

undoAnimDuration :: Float

Duração da animação do movimento do jogador e das caixas quando se anula uma jogada.

restartAnimDuration :: Float

Duração da animação de cada movimento do jogador e das caixas quando se anula todas as jogadas já feitas.