Sokoban.Game.Level

Description

Este módulo gere todos os aspetos relativos à representação e interação do utilizador com o estado de um nível.

Tipos

data Level

Tipo abstrato que representa um nível.

Constructors

Level

Fields

levelState :: LevelState

Estado do nível.

levelReplayCmds :: [Command]

Comandos que devem ser executados automaticamente no início do jogo.

levelBoard :: Board

Tabuleiro.

levelPlayerAttr :: PlayerAttr

Atributos do personagem.

levelPlayer :: Player

Jogador.

levelBoxes :: Boxes

Caixas.

levelAnimTime :: Float

Tempo atual da animação.

levelAnimDuration :: Float

Duração da animação.

levelMoveLeft :: Bool

Determina se o utilizador pretende que o jogador se mova para a esquerda.

levelMoveDown :: Bool

Determina se o utilizador pretende que o jogador se mova para baixo.

levelMoveRight :: Bool

Determina se o utilizador pretende que o jogador se mova para a direita.

levelMoveUp :: Bool

Determina se o utilizador pretende que o jogador se mova para cima.

levelKeySpace :: Bool

Determina se a tecla de espaço está premida.

data LevelState

Representa o estado de um nível.

Constructors

Playing
O nível não está terminado e o utilizador pode interagir com o personagem.

Replaying
O nível está a executar comandos automáticos.

Restarting
O nível está anular todos os movimentos do jogador de modo a regressar ao estado inicial.

Ended
O nível terminou, quer por o jogador ter arrumado todas as caixas, quer por desistência.

type Command = Char

Representa um comando que deve ser executado automaticamente no início do nível. Estes comandos descrevem movimentos do jogador.

Construção

create

Arguments

:: Assets
Assets.

-> String
Nível a ser jogado.

-> PlayerAttr
Atributos do personagem a utilizar.

-> IO Level

Contrói um valor do tipo Level, dada uma String que representa o estado inicial do nível e os atributos do personagem a utilizar.

Parsing

parseLevel

Arguments

:: Assets
Assets.

-> Theme
Tema.

-> PlayerAttr
Atributos do personagem a utilizar.

-> String
Nível a ser jogado.

-> IO (Board, Player, Boxes, [Command])

Contrói valores do tipo Board, Player, Boxes e [Command] a partir de uma representação textual adequada.

parseCells :: [String] -> ([Cell], VectorI)

Contrói as células de um tabuleiro a partir de uma representação textual adequada.

parseCoords :: [String] -> [VectorI]

Contrói uma lista de pontos a partir de uma representação textual adequada.

parsePos :: String -> VectorI

Contrói um ponto a partir de uma representação textual adequada.

parseCommands :: [String] -> [Command]

Contrói uma lista de comandos a partir de uma representação textual adequada.

Estado

Funções que inspecionam o estado de um nível.

isAnimating :: Level -> Bool

Verifica se o nível se encontra a meio de uma animação.

numSteps :: Level -> Int

Determina o número de movimentos que o jogador já efetuou.

levelFinished :: Level -> Bool

Verifica se o tabuleiro e as caixas de um nível se encontram num estado terminal, i.e. se o nível já foi terminado.

Ações

Funções exportadas que executam ações sobre um nível, e.g. anular uma jogada.

undo :: Level -> Level

Anula a última ação realizada pelo jogador, se existente.

restart :: Level -> Level

Anula todas as ações já realizadas pelo jogador.

giveUp :: Level -> Level

Termina o jogo por desistência.

Movimento do Jogador

Funções responsáveis por gerir o movimento do jogador.

playerMove

Arguments

:: VectorI
Direção do movimento.

-> Level

-> Level

Tenta mover o jogador numa certa direção.

playerMoveFree

Arguments

:: VectorI
Direção do movimento.

-> Level

-> Level

Tenta mover o jogador numa certa direção, no caso de o jogador não se encontrar a agarrar uma caixa.

playerMoveGrabbing

Arguments

:: VectorI
Direção do movimento.

-> Level

-> Level

Tenta mover o jogador numa certa direção, no caso de o jogador se encontrar a agarrar uma caixa.

playerCanMove

Arguments

:: Int
Número de caixas que o jogador ainda pode empurrar.

-> VectorI
Posição para onde o jogador se deve mover.

-> VectorI
Direção na qual o jogador se deve mover.

-> Level

-> Maybe [VectorI]

Determina se o jogador se pode mover numa certa direção, tendo em conta o número de caixas que este pode empurrar.

Se o jogador não se puder mover, é devolvido o valor Nothing. Em caso contrário, é devolvido o valor Just boxes, onde boxes corresponde às posições de todas as caixas que o jogador iria movimentar ao realizar o movimento em questão.

playerGrab :: VectorI -> Level -> Level

Faz com que o jogador agarre na caixa que se encontra na direção especificada, se possível.

playerSaveTeleport :: Level -> Level

Armazena a posição atual do jogador como a posição de destino do teleporte.

playerDoTeleport :: Level -> Level

Teleporta o jogador para a posição previamente armazenada, se possível.

playerCanTeleport :: Level -> Bool

Verifica se o jogador se pode teleportar para a posição previamente armazenada.

Processamento de Eventos

handleEvent

Arguments

:: Event
Evento.

-> Level

-> Level

Processa um evento.

handleDirKey

Arguments

:: SpecialKey
Tecla premida.

-> Level

-> Level

Processa um evento relativo a uma tecla de direção.

Atualização

update

Arguments

:: Float
Tempo desde a última atualização, em segundos.

-> Level

-> (Level, Bool)

Atualiza o estado de um nível.

É devolvido um tuplo com o novo estado do nível e um booleano que determina se durante esta atualização se verificou, pela primeira vez, que o nível terminou.

updatePlaying :: Level -> Level

Função auxiliar da função update, utilizada quando o estado do nível tem o valor Playing.

updateReplaying :: Level -> Level

Função auxiliar da função update, utilizada quando o estado do nível tem o valor Replaying.

updateRestarting :: Level -> Level

Função auxiliar da função update, utilizada quando o estado do nível tem o valor Restarting.

Conversão para Picture

drawingRegion :: Level -> Rect

Determina a região em que um nível será desenhado.

draw :: Level -> Picture

Desenha um nível.