Sokoban.Game.Player

Description

Este módulo gere o armazenamento e representação visual das personagens que o utilizador controla ao jogar o jogo.

Tipos

data Player

Tipo abstrato que representa um jogador.

Constructors

Player

Fields

playerPos :: VectorI

Coordenadas da célula do tabuleiro onde o jogador se encontra.

playerDir :: VectorI

Direção para onde o jogador está virado.

playerIsGrabbing :: Bool

Determina se o jogador está a agarrar uma caixa.

playerTeleportPos :: Maybe VectorI

Posição de destino do teleporte.

playerFinished :: Bool

Determina se o jogador terminou o jogo.

playerAnimPos :: Vector

Posição visual atual do jogador.

playerAnimInitPos :: Vector

Posição visual do jogador no início da animação.

playerAnimFinalPos :: Vector

Posição visual do jogador no final da animação.

playerAnimPhase :: Int

Fase atual da animação. Toma o valor 0 ou 1.

playerPicFun :: String -> Picture

Função que obtém uma imagem do jogador dado o identificador de um bitmap.

playerPic :: Picture

Imagem atual do jogador.

playerSteps :: [Step]

Movimentos que o jogador já realizou.

data Step

Armazena informação sobre um movimento do jogador.

Constructors

StepMovement

Fields

stepInvDir :: VectorI

Direção oposta á do movimento.

stepTurnDir :: VectorI

Direção para o qual o jogador estava virado durante o movimento.

stepIsGrabbing :: Bool

Determina se o jogador estava a agarrar numa caixa durante o movimento.

stepTeleportPos :: Maybe VectorI

Posição de destino do teleporte antes do movimento.

StepTeleport

Fields

stepPrevPos :: VectorI

Posição do jogador antes do teleporte.

Construção

create

Arguments

:: Assets
Assets.

-> PlayerAttr
Atributos do personagem.

-> VectorI
Posição inicial no tabuleiro.

-> Player

Contrói um valor do tipo Player, dados os atributos do personagem e a posição inicial do jogador.

Estado

pos :: Player -> VectorI

Determina as coordenadas da célula do tabuleiro onde um jogador se encontra.

dir :: Player -> VectorI

Determina a direção para onde um jogador está virado.

isGrabbing :: Player -> Bool

Verifica se um jogador está a agarrar uma caixa.

teleportPos :: Player -> Maybe VectorI

Determina a posição de destino de teleporte de um jogador.

numSteps :: Player -> Int

Determina o número de movimentos que o jogador já efetuou.

canUndo :: Player -> Bool

Verifica se o jogador pode anular uma jogada, i.e., se o jogador tem jogadas para anular.

Ações

move :: VectorI -> Player -> Player

Move o jogador uma célula na direção especificada.

setDir

Arguments

:: VectorI
Direção para onde o jogador deve ficar virado.

-> Player

-> Player

Altera a direção para onde um jogador está virado.

grabBox :: Player -> Player

Faz com que o jogador agarre uma caixa, na direção para onde ele se encontra virado.

releaseBox :: Player -> Player

Faz com que o jogador deixe de agarrar uma caixa.

saveTeleport :: Player -> Player

Armazena a posição atual do jogador como a sua posição de teleporte.

doTeleport :: Player -> Player

Teleporta o jogador para a posição previamente armazenada.

finish :: Player -> Player

Altera o aspeto do jogador para uma imagem que indique que o nível foi terminado.

undo :: Player -> Player

Anula o último movimento do jogador.

undoStep

Arguments

:: Step
Movimento a anular.

-> Player

-> Player

Função auxiliar utilizada pela função undo, que anula um movimento do jogador.

Atualização

update

Arguments

:: Float
Fator, de 0 a 1, que determina o estado da animação atual.

-> Player

-> Player

Atualiza o estado de um jogador.

Conversão para Picture

draw :: Player -> Picture

Desenha uma jogador.