Sokoban.Game.PlayerAttr

Description

Este módulo define um tipo que representa as capacidades (atributos) de um personagem, e.g. a força do personagem. Também disponibilizia uma lista com as capacidades das personagens disponíveis no jogo.

Documentation

data PlayerAttr

Representa as capacidades e outras informações sobre um personagem.

Constructors

PlayerAttr

Fields

name :: String

Nome do personagem.

description :: String

Descrição do personagem.

pushStrength :: Int

Número de caixas que o personagem pode empurrar simultaneamente.

canPull :: Bool

Determina se o personagem pode puxar caixas.

canTeleport :: Bool

Determina se o personagem se pode teleportar.

bitmapName :: String

Identificador dos bitmaps do personagem.

attributes :: [PlayerAttr]

Atributos dos personagens disponíveis no jogo.