Sokoban.Game.Stats

Description

Este módulo gere o painel de informações sobre o jogo, que apresenta o número do nível atual, o número total de níveis, o número de passos já efetuados e o tempo decorrido desde o início do nível.

Documentation

backgroundCenter :: Vector

Centro da região de desenho da imagem de fundo.

backgroundRegion :: Rect

Região de desenho da imagem de fundo.

backgroundWidth :: Float

Largura da região de desenho da imagem de fundo.

backgroundHeight :: Float

Altura da região de desenho da imagem de fundo.

backgroundSpacing :: Float

Tamanho das margens entre o espaço de desenho da imagem de fundo e o espaço de desenho do painel de informação.

textHeight :: Float

Altura do texto do painel de informação.

data Stats

Tipo abstrato que representa um painel de informação.

Constructors

Stats

Fields

statsBackPic :: Picture

Imagem de fundo do painel.

statsTextPic :: Picture

Imagem do texto do painel.

statsLevelCur :: Int

Número do nível atual.

statsLevelNum :: Int

Número total de níveis.

statsSteps :: Int

Número de passos.

statsTimeInit :: TimeSpec

Instante de tempo no início do nível atual.

statsTimeSec :: Int

Número de segundos decorridos desde o início do nível atual.

statsTimeFrozen :: Bool

Determina se o temporizador está parado.

create

Arguments

:: Assets
Assets.

-> Int
Número total de níveis.

-> IO Stats

Contrói um painel de informação sobre o jogo, dado o número total de níveis disponíveis para serem jogados.

createTextPic

Arguments

:: Assets
Assets.

-> Int
Número do nível atual.

-> Int
Número total de níveis.

-> Int
Número de passos.

-> Int
Número de segundos decorridos desde o início do nível atual.

-> Picture

Contrói a imagem do texto de um painel de informação.

secondsToString :: Int -> String

Converte um número de segundos numa representação textual adequada.

Exemplos de utilização:

>>> secondsToString 0
"00:00"

>>> secondsToString 42
"00:42"

>>> secondsToString 1672
"27:52"

advanceLevel

Arguments

:: Assets
Assets.

-> Stats

-> IO Stats

Incrementa o número do nível atual e anula o número de passos e o temporizador de um painel de informação.

setSteps

Arguments

:: Assets
Assets.

-> Int
Número de passos.

-> Stats

-> Stats

Define o número de passos de um painel de informação.

freezeTimer :: Stats -> Stats

Bloqueia o temporizador de um painel de informação.

update

Arguments

:: Assets
Assets.

-> Stats
Tempo desde a última atualização, em segundos.

-> IO Stats

Atualiza o estado de um painel de informação.

drawingRegion :: Rect

Determina a região em que um painel de informação será desenhado.

draw :: Stats -> Picture

Desenha um painel de informação.