Sokoban.Helper

Description

Este módulo contém tipos e funções úteis a vários fins.

Coleções

insertWithAL

Arguments

:: Eq k

=> (a -> a -> a)
Função utilizada quando já existe um elemento com a mesma chave.

-> k
Chave do elemento a inserir.

-> a
Valor do elemento a inserir.

-> [(k, a)]
Lista na qual inserir o elemento.

-> [(k, a)]
Lista resultante.

Insere um valor com a chave especificada numa lista de associação.

Se já existir um valor com a mesma chave na lista, esse valor é substituido por:

joinFun oldValue newValue

joinFun - a função especificada pelo utilizador;

oldValue - o valor já associado com a chave;

newValue - o valor que o utilizador tentou inserir.

Exemplos de utilização:

>>> insertWithAL (\x _ -> x) 2 'b' []
[(2,'b')]

>>> insertWithAL (\x _ -> x) 2 'b' [(3,'a')]
[(3,'a'),(2,'b')]

>>> insertWithAL (\x _ -> x) 2 'b' [(2,'a')]
[(2,'b')]

>>> insertWithAL (\_ y -> y) 2 'b' [(2,'a')]
[(2,'a')]

Matemática

type VectorI = (Int, Int)

Representa um vetor bidimensional de componentes do tipo Int.

fromIntegralVector :: (Integral a, Num b) => (a, a) -> (b, b)

Converte um vetor bidimensional de componentes com tipo na classe Integral para um vetor bidimensional de componentes com tipo na classe Num.

Exemplos de utilização:

>>> fromIntegralVector ((1,2) :: (Int, Int))
(1,2) :: (Num a) => (a, a)

(|+|) :: Num a => (a, a) -> (a, a) -> (a, a)

Adiciona dois vetores bidimensionais componente por componente, segundo a equação:

(x1, y1) |+| (x2, y2) = (x1+x2, y1+y2)

Exemplos de utilização:

>>> (1,4) |+| (2,5)
(3,9)

(|-|) :: Num a => (a, a) -> (a, a) -> (a, a)

Subtrai dois vetores bidimensionais componente por componente, segundo a equação:

(x1, y1) |-| (x2, y2) = (x1-x2, y1-y2)

Exemplos de utilização:

>>> (2,7) |-| (1,5)
(1,2)

vecNeg :: Num a => (a, a) -> (a, a)

Nega os componentes de um vetor bidimensional, segundo a equação:

vecNeg (x,y) = (-x, -y)

Exemplos de utilização:

>>> vecNeg (1,2)
(-1,-2)

(*|) :: Num a => a -> (a, a) -> (a, a)

Multiplica um número por um vetor bidimensional, segundo a equação:

s *| (x,y) = (s*x, s*y)

Exemplos de utilização:

>>> 3 *| (1,2)
(3,6)

(|*|) :: Num a => (a, a) -> (a, a) -> (a, a)

Multiplica dois vetores bidimensionais componente por componente, segundo a equação:

(x1, y1) |*| (x2, y2) = (x1*x2, y1*y2)

Exemplos de utilização:

>>> (1,3) |*| (2,3)
(2,9)

lerp1 :: Num a => a -> a -> a -> a

Interpolação linear em uma dimensão, segundo a equação:

lerp1 t p0 p1 = p0 + (p1-p0) * t

Exemplos de utilização:

>>> lerp1 0 2 4
2

>>> lerp1 0.5 2 4
3

>>> lerp1 1 2 4
4

lerp2 :: Num a => a -> (a, a) -> (a, a) -> (a, a)

Interpolação linear em duas dimensões, segundo a equação:

lerp2 t (x0,y0) (x1,y1) = (lerp1 t x0 x1, lerp1 t y0 y1)

Exemplos de utilização:

>>> lerp2 0 (2,4) (4,8)
(2,4)

>>> lerp2 0.5 (2,4) (4,8)
(3,6)

>>> lerp2 1 (2,4) (4,8)
(4,8)

data Rect

Representa um retângulo alinhado com os eixos.

Constructors

Rect

Fields

rectLeft :: Float

Coordenada, ao longo do eixo Ox, da aresta esquerda do retângulo.

rectBottom :: Float

Coordenada, ao longo do eixo Oy, da aresta inferior do retângulo.

rectRight :: Float

Coordenada, ao longo do eixo Ox, da aresta direita do retângulo.

rectTop :: Float

Coordenada, ao longo do eixo Oy, da aresta superior do retângulo.

makeRect :: Vector -> Rect

Contrói um retângulo alinhado com os eixos centrado na origem e com o tamanho especificado.

Exemplos de utilização:

>>> makeRect (4,2)
Rect (-2) (-1) 2 1

makeRectAt :: Vector -> Vector -> Rect

Contrói um retângulo alinhado com os eixos centrado no ponto e com o tamanho especificados.

Exemplos de utilização:

>>> makeRectAt (3,3) (4,2)
Rect 1 2 5 4

rectCenter :: Rect -> Vector

Calcula as coordenadas do centro de um retângulo representado por um valor do tipo Rect.

Exemplos de utilização:

>>> rectCenter (Rect 1 2 5 4)
(3,3)

rectSize :: Rect -> Vector

Calcula as dimensões de um retângulo representado por um valor do tipo Rect.

Exemplos de utilização:

>>> rectSize (Rect 1 5 4 7)
(3,2)

rectWidth :: Rect -> Float

Calcula a largura de um retângulo representado por um valor do tipo Rect.

Exemplos de utilização:

>>> rectWidth (Rect 1 5 4 7)
3

rectHeight :: Rect -> Float

Calcula a altura de um retângulo representado por um valor do tipo Rect.

Exemplos de utilização:

>>> rectHeight (Rect 1 5 4 7)
2

pointIsInsideRect :: Vector -> Rect -> Bool

Verifica se um ponto se encontra dentro de um retângulo representado por um valor do tipo Rect.

Exemplos de utilização:

>>> (0,0) `pointIsInsideRect` (Rect 1 5 4 7)
False

>>> (3,5) `pointIsInsideRect` (Rect 1 5 4 7)
True

Gráficos

bitmapSize :: Picture -> Vector

Devolve o tamanho de um bitmap representado por um valor do tipo Picture construído pelo construtor Bitmap.

type RectMapping = Rect -> Rect

Representa uma transformação entre valores do tipo Rect.

type PictureMapping = Picture -> Picture

Representa uma transformação entre valores do tipo Picture.

regionMapX

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Float
Coordenada relativa à região original.

-> Float
Coordenada relativa à região de destino.

Transforma uma coordenada ao longo do eixo Ox, relativa a uma região retangular, de maneira a que esta passe a ser relativa a outra região retangular.

Exemplos de utilização:

>>> regionMapX (Rect 1 3 2 7) (Rect 4 5 0 2) 0
8

>>> regionMapX (Rect 1 3 2 7) (Rect 4 5 0 2) 3
-4

regionMapY

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Float
Coordenada relativa à região original.

-> Float
Coordenada relativa à região de destino.

Transforma uma coordenada ao longo do eixo Oy, relativa a uma região retangular, de maneira a que esta passe a ser relativa a outra região retangular.

Exemplos de utilização:

>>> regionMapY (Rect 1 3 2 7) (Rect 4 5 0 2) 0
7.25

>>> regionMapY (Rect 1 3 2 7) (Rect 4 5 0 2) 3
5

regionMapRect

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Rect
Retângulo relativo à região original.

-> Rect
Retângulo relativo à região de destino.

Transforma um retângulo, relativo a uma região retangular, de maneira a que este passe a ser relativo a outra região retangular.

Exemplos de utilização:

>>> regionMapRect (Rect 1 3 2 7) (Rect 4 5 0 2) (Rect 1 2 1 4)
Rect 4 5.75 4 4.25

>>> regionMapRect (Rect 1 3 2 7) (Rect 4 5 0 2) (Rect 0 0 1 1)
Rect 8 7.25 4 6.5

regionMapPicture

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Picture
Picture relativa à região original.

-> Picture
Picture relativa à região de destino.

Transforma uma Picture, relativa a uma região retangular, de maneira a que esta passe a ser relativa a outra região retangular.

regionMapBitmap

Arguments

:: Rect
Região que o bitmap passa a ocupar.

-> Picture
Picture a transformar.

-> Picture
Picture transformada.

Transforma uma Picture do construtor Bitmap de maneira a que esta passe a ocupar a região retangular especificada.

zoomRegionToTop

Arguments

:: Rect
Região a redimensionar.

-> Rect
Região de destino.

-> Rect
Região resultante.

Redimensiona uma região retangular, mantendo o seu aspect ratio, de maneira a ocupar o máximo de espaço de outra região retangular. A região resultante fica, em relação à região de destino, centrada horizontalmente e encostada acima.

Exemplos de utilização:

>>> zoomRegionToTop (Rect 1 3 2 7) (Rect 4 5 0 2)
Rect 4 18 0 2

zoomRegionToCenter

Arguments

:: Rect
Região a redimensionar.

-> Rect
Região de destino.

-> Rect
Região resultante.

Redimensiona uma região retangular, mantendo o seu aspect ratio, de maneira a ocupar o máximo de espaço de outra região retangular. A região resultante fica, em relação à região de destino, centrada horizontalmente e verticalmente.

Exemplos de utilização:

>>> zoomRegionToCenter (Rect 1 3 2 7) (Rect 4 5 0 2)
Rect 4 11.5 0 (-4.5)

zoomRegionToBottom

Arguments

:: Rect
Região a redimensionar.

-> Rect
Região de destino.

-> Rect
Região resultante.

Redimensiona uma região retangular, mantendo o seu aspect ratio, de maneira a ocupar o máximo de espaço de outra região retangular. A região resultante fica, em relação à região de destino, centrada horizontalmente e encostada abaixo.

Exemplos de utilização:

>>> zoomRegionToBottom (Rect 1 3 2 7) (Rect 4 5 0 2)
Rect 4 5 0 (-11)

zoomRectToTop

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Rect
Retângulo relativo à região original.

-> Rect
Retângulo relativo à região resultante.

Combinação das funções zoomRegionToTop e regionMapRect, definida por:

zoomRectToTop src dst r = regionMapRect src (zoomRegionToTop src dst) r

Exemplos de utilização:

>>> zoomRectToTop (Rect 1 3 2 7) (Rect 4 5 0 2) (Rect 1 2 1 4)
Rect 4 22 4 14

zoomRectToCenter

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Rect
Retângulo relativo à região original.

-> Rect
Retângulo relativo à região resultante.

Combinação das funções zoomRegionToCenter e regionMapRect, definida por:

zoomRectToCenter src dst r = regionMapRect src (zoomRegionToCenter src dst) r

Exemplos de utilização:

>>> zoomRectToCenter (Rect 1 3 2 7) (Rect 4 5 0 2) (Rect 1 2 1 4)
Rect 4 15.5 4 7.5

zoomRectToBottom

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Rect
Retângulo relativo à região original.

-> Rect
Retângulo relativo à região resultante.

Combinação das funções zoomRegionToBottom e regionMapRect, definida por:

zoomRectToBottom src dst r = regionMapRect src (zoomRegionToBottom src dst) r

Exemplos de utilização:

>>> zoomRectToBottom (Rect 1 3 2 7) (Rect 4 5 0 2) (Rect 1 2 1 4)
Rect 4 9 4 1

zoomPictureToTop

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Picture
Picture relativa à região original.

-> Picture
Picture relativa à região resultante.

Combinação das funções zoomRegionToTop e regionMapPicture, definida por:

zoomPictureToTop src dst p = regionMapPicture src (zoomRegionToTop src dst) p

zoomPictureToCenter

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Picture
Picture relativa à região original.

-> Picture
Picture relativa à região resultante.

Combinação das funções zoomRegionToCenter e regionMapPicture, definida por:

zoomPictureToCenter src dst p = regionMapPicture src (zoomRegionToCenter src dst) p

zoomPictureToBottom

Arguments

:: Rect
Região original.

-> Rect
Região de destino.

-> Picture
Picture relativa à região original.

-> Picture
Picture relativa à região resultante.

Combinação das funções zoomRegionToBottom e regionMapPicture, definida por:

zoomPictureToBottom src dst p = regionMapPicture src (zoomRegionToBottom src dst) p

rectWirePicture :: Rect -> Picture

Contrói um valor do tipo Picture que representa as arestas do retângulo especificado.

multBMPToPicture

Arguments

:: Color
Cor pela qual multiplicar o bitmap.

-> BMP
Bitmap a alterar.

-> Picture
Bitmap resultante.

Multiplica um bitmap representado por um valor do tipo BMP pela cor especificada e converte o resultado para um valor do tipo Picture.

convertRGBAToABGR :: [Float] -> [Float]

Converte uma lista de valores do tipo Float que representa uma série de píxeis cujos componentes têm a ordem RGBA para uma lista equivalente com os componentes na ordem ABGR.

multPixels

Arguments

:: (Float, Float, Float, Float)
Cor pela qual multiplicar os píxeis (RGBA).

-> [Float]
Componentes dos píxeis a alterar.

-> [Float]
Componentes dos píxeis resultantes.

Multiplica uma lista de píxeis (com componentes na ordem RGBA) pela cor especificada.

Input do Utilizador

normalizeKey :: Key -> Key

Convert valores da forma Char c em Char (toLower c). Valores de outras formas permanecem iguais.

Exemplos de utilização:

>>> normalizeKey (Char 'A')
Char 'a'

>>> normalizeKey (Char 'a')
Char 'a'

>>> normalizeKey (SpecialKey KeyEnter)
SpecialKey KeyEnter

keyToString :: Key -> String

Convert um valor do tipo Key numa String que o identifique.

Exemplos de utilização:

>>> keyToString (Char 'A')
"A"

>>> keyToString (Char 'a')
"A"

>>> keyToString (SpecialKey KeyEsc)
"Esc"

>>> keyToString (MouseButton LeftButton)
"LeftButton"

Tempo

secondsBetween :: TimeSpec -> TimeSpec -> Int

Calcula o número de segundos (arredondado por defeito) entre dois instantes de tempo determinados por valores do tipo TimeSpec.

Sistema de Ficheiros

fileName :: FilePath -> FilePath

Extrai o nome do ficheiro de um caminho para um ficheiro.

Exemplos de utilização:

>>> fileName ""
""

>>> fileName "dir1/dir2/"
""

>>> fileName "dir1/dir2/ficheiro"
"ficheiro"

>>> fileName "dir1/dir2/ficheiro.ext"
"ficheiro.ext"

fileNameWithoutExt :: FilePath -> FilePath

Extrai o nome do ficheiro sem extensão de um caminho para um ficheiro.

Exemplos de utilização:

>>> fileNameWithoutExt ""
""

>>> fileNameWithoutExt "dir1/dir2/"
""

>>> fileNameWithoutExt "dir1/dir2/ficheiro"
"ficheiro"

>>> fileNameWithoutExt "dir1/dir2/ficheiro.ext"
"ficheiro"

removeExt :: FilePath -> FilePath

Remove a extensão de um caminho para um ficheiro.

Exemplos de utilização:

>>> removeExt ""
""

>>> removeExt "dir1/dir2/"
"dir1/dir2/"

>>> removeExt "dir1/dir2/ficheiro"
"dir1/dir2/ficheiro"

>>> removeExt "dir1/dir2/ficheiro.ext"
"dir1/dir2/ficheiro"

dirName :: FilePath -> FilePath

Extrai o nome da diretoria de um caminho para uma diretoria.

O caminho especificado deve terminar com '/'.

Exemplos de utilização:

>>> dirName "/"
""

>>> dirName "dir1/dir2/"
"dir2"

getFilesInDir :: FilePath -> IO [FilePath]

Devolve uma lista com os caminhos para os ficheiros na diretoria especificada.

O caminho especificado deve terminar com '/'.

getDirsInDir :: FilePath -> IO [FilePath]

Devolve uma lista com os caminhos para as subdiretorias na diretoria especificada.

O caminho especificado deve terminar com '/'.

getDirContents :: FilePath -> IO [FilePath]

Devolve uma lista com os caminhos para os ficheiros e subdiretorias na diretoria especificada.

O caminho especificado deve terminar com '/'.