Sokoban.Menu.Menu

Description

Este módulo gere o menu principal, através do qual o jogador pode selecionar o personagem.

Documentation

data Menu

Tipo abstrato que representa um menu.

Constructors

Menu

Fields

menuTitlePic :: Picture

Imagem do título.

menuSelector :: PlayerSelector

Selecionador de personagem.

menuPlayButton :: Button

Botão para jogar.

menuQuitButton :: Button

Botão para sair.

menuRegion :: Rect

Região de desenho do menu.

makeRectMapping :: Shared -> RectMapping

Gera uma transformação de retângulos entre espaço do menu e espaço da janela.

makePictureMapping :: Shared -> PictureMapping

Gera uma transformação de Pictures entre espaço do menu e espaço da janela.

create

Arguments

:: Shared
Estado partilhado.

-> Menu

Cria uma instância do menu principal.

handleEvent

Arguments

:: Shared
Estado partilhado.

-> Event
Evento.

-> Menu
Menu.

-> IO (Either Menu PlayerAttr)

Processa um comando.

Se o menu deve continuar a existir, devolve Left newMenu, onde newMenu corresponde ao novo estado do menu.

Se o menu tiver terminado, devolve Right attr, onde attr corresponde aos atributos do personagem selecionado.

update

Arguments

:: Shared
Estado partilhado.

-> Float
Tempo desde a última atualização, em segundos.

-> Menu
Menu.

-> Menu

Atualiza o estado do menu.

draw

Arguments

:: Shared
Estado partilhado.

-> Menu
Menu.

-> Picture

Desenha um menu.