Sokoban.Menu.PlayerSelector

Description

Este módulo permite a utilização de um selecionador de personagem interativo na interface gráfica do utilizador.

Documentation

itemSize :: Vector

Dimensões de cada visualizador de personagem.

itemSpacing :: Float

Espaço entre cada visualizador de personagem.

itemAnimDuration :: Float

Duração da animação do personagem selecionado.

selectionColor :: Color

Cor do retângulo envolvente da visualizador de personagem selecionada.

mouseSelectionColor :: Color

Cor do retângulo envolvente da visualizador de personagem em que o cursor se encontra.

data PlayerSelector

Tipo abstrato que representa um selecionador de personagem.

Constructors

PlayerSelector

Fields

selectorTextPic :: Picture

Imagem do texto de cabeçalho.

selectorItems :: [Item]

Visualizadores dos personagens.

selectorSelection :: VectorI

Coordenadas do visualizador atualmente selecionado.

selectorMouseSelection :: Maybe VectorI

Coordenadas do visualizador sobre o qual o cursor se encontra.

data Item

Representa a área de um personagem.

Constructors

Item

Fields

itemCoords :: VectorI

Coordenadas que identificam o visualizador.

itemRegion :: Rect

Região que o visualizador ocupa.

itemAnimTime :: Float

Tempo atual de animação.

itemPicFun :: Float -> Picture

Função que obtém a imagem do personagem dado um determinado tempo de animação.

itemPic :: Picture

Imagem atual do personagem.

itemText :: Picture

Imagem do texto do visualizador.

selection

Arguments

:: PlayerSelector
O selecionador de personagem.

-> PlayerAttr
Os atributos correspondentes ao personagem selecionado.

Seleção atual de um selecionador de personagem.

coordsIndex :: VectorI -> Int

Faz corresponder um índice a um par de coordenadas que identificam um visualizador.

mouseSelection

Arguments

:: PlayerSelector
Selecionador de personagem.

-> Vector
Posição do cursor no espaço da janela.

-> RectMapping
Transformação entre espaço do selecionador e espaço da janela.

-> Maybe VectorI
Coordenadas do visualizador.

Verifica sobre que visualizador o cursor se encontra.

create

Arguments

:: Assets
Assets.

-> Vector
Posição do centro do selecionador.

-> PlayerSelector

Contrói um selecionador de personagem.

createItem

Arguments

:: Assets
Assets.

-> (VectorI, Rect, PlayerAttr)
(Coordenadas, Região, Atributos do personagem)

-> Item

Contrói um visualizador.

createItemPic

Arguments

:: (String -> Picture)
Função que obtém uma imagem do personagem dado o nome dessa imagem.

-> Vector
Posição do centro da imagem.

-> Float
Tempo de animação.

-> Picture
Imagem do personagem.

Obtém a imagem de um personagem dado um tempo de animação.

handleEvent

Arguments

:: Event
Evento.

-> RectMapping
Transformação entre espaço do selecionador e espaço da janela.

-> PlayerSelector
Selecionador de personagem.

-> PlayerSelector

Processa um evento.

update

Arguments

:: Float
Tempo desde a última atualização, em segundos.

-> PlayerSelector
Selecionador de personagem.

-> PlayerSelector

Atualiza o estado de um selecionador de personagem.

updateItem

Arguments

:: Float
Tempo desde a última atualização, em segundos.

-> VectorI
Seleção atual.

-> Item
Visualizador.

-> Item

Atualiza o estado de um visualizador.

draw :: PlayerSelector -> Picture

Desenha um selecionador de personagem.

drawItem :: Item -> [Picture]

Desenha um visualizador.