Sokoban.Shared

Description

Este módulo contém um tipo e funções destinados a facilitar a partilha de certas estruturas de dados cuja utilização é recorrente em várias partes do programa.

Documentation

data Shared

Tipo abstrato que armazena estruturas de dados que são frequentemente utilizadas por várias partes do programa.

Constructors

Shared

Fields

assets :: Assets

window :: Window

create :: VectorI -> IO Shared

Cria um valor do tipo Shared, cuja Window tem a resolução inicial especificada.

handleEvent

Arguments

:: Event
Evento.

-> Shared

-> Shared

Processa um evento.