Sokoban.UI.Button

Description

Este módulo permite a utilização botões interativos na interface gráfica do utilizador.

Documentation

Um botão é constituído por uma imagem de fundo comum a todos os botões e algum texto que identifica o botão. O utilizador pode interagir com o botão utilizando o rato ou carregando na tecla associada a esse botão.

data Button

Tipo abstrato que representa um botão.

Constructors

Button

Fields

buttonRegion :: Rect

Região onde o botão será desenhado.

buttonBackPicUp :: Picture

Imagem do botão quando o cursor se encontra fora da região do botão.

buttonBackPicOver :: Picture

Imagem do botão quando o cursor se encontra dentro da região do botão.

buttonBackPicDown :: Picture

Imagem do botão quando o cursor se encontra dentro da região do botão e o botão esquerdo do rato está premido.

buttonTextPic :: Picture

Texto do botão.

buttonKey :: Key

Tecla associada com o botão.

buttonState :: ButtonState

Estado a atual do botão.

buttonMouseDown :: Bool

Define se o botão esquerdo do rato foi pressionado dentro da sua região.

data ButtonState

Representa o estado de um botão.

Constructors

ButtonStateUp
O cursor encontra-se fora da região do botão.

ButtonStateOver
O cursor encontra-se dentro da região do botão e o botão esquerdo do rato não está premido.

ButtonStateDown
O cursor encontra-se dentro da região do botão e o botão esquerdo do rato está premido.

create

Arguments

:: Assets
Assets.

-> Rect
Região onde o botão será desenhado.

-> String
Texto do botão.

-> Key
Tecla associada com o botão.

-> Button
Botão resultante.

Contrói um botão.

handleEvent

Arguments

:: Event
Evento.

-> RectMapping
Transformação entre espaço do botão e espaço da janela.

-> Button
Botão.

-> (Button, Bool)

Processa um evento.

É devolvido um tuplo com o novo botão e um booleano que indica se o botão acabou de ser premido.

pointIsInside

Arguments

:: Vector
Ponto.

-> RectMapping
Transformação entre espaço do botão e espaço da janela.

-> Button
Botão.

-> Bool

Verifica se um ponto se encontra dentro da região de um botão, tendo em conta a transformação especificada.

draw :: Button -> Picture

Desenha um botão.