Sokoban.UI.Font

Description

Este módulo permite desenhar texto com formatação e tipos de letra arbitrários, carregados a partir de uma série de ficheiro num formato adequado.

Tipos

data Text

Representa uma secção de texto formatado.

Constructors

Text

Fields

textString :: String

Texto.

textHeight :: Float

Altura de uma linha de texto.

textAlignment :: Alignment

Alinhamento do texto.

textColor :: Color

Cor do texto.

textShadowColor :: Maybe Color

Cor da sombra. Nothing para não desenhar sombra.

data Alignment

Define o alinhamento de uma secção de texto.

Constructors

LeftTop
Encostado à esquerda e acima.

CenterTop
Centrado horizontalmente e encostado acima.

RightTop
Encostado à direita e acima.

LeftCenter
Encostado à esquerda e centrado verticalmente.

Center
Centrado horizontal e verticalmente.

RightCenter
Encostado à direita e centrado verticalmente.

LeftBottom
Encostado à esquerda e abaixo.

CenterBottom
Centrado horizontalmente e encostado abaixo.

RightBottom
Encostado à direita e abaixo.

data FontConfig

Tipo utilizado pelo ficheiro de configuração do tipo de letra.

Constructors

FontConfig

Fields

height :: Float

spacing :: Float

globalOffsetY :: Float

newlineOffset :: Float

shadowOffset :: Maybe Vector

data Font

Representa um tipo de letra.

Constructors

Font

Fields

fontHeight :: Float

Altura de uma linha de texo.

fontSpacing :: Float

Espaço horizontal entre carateres.

fontYOffset :: Float

Ajuste vertical global.

fontNewlineOffset :: Float

Distância entre linhas.

fontShadowOffset :: Maybe Vector

Diferença de posição da sombra.

fontChars :: Map Char FontChar

Mapa de carateres suportados pelo tipo de letra.

data FontChar

Representa um caratere de um tipo de letra.

Constructors

FontChar

Fields

fontCharSize :: Vector

Dimensões do caratere.

fontCharYOffset :: Float

Ajuste vertical na posição do caratere.

fontCharPics :: [(Color, Picture)]

Bitmaps do caratere, em várias cores.

Carregamento

load

Arguments

:: FilePath
Caminho da diretoria.

-> [Color]
Cores suportadas.

-> IO Font

Carrega um tipo de letra a partir de uma diretoria. Essa diretoria deve conter um ficheiro de configuração adequado e imagens do tipo BMP que representem os carateres do tipo de letra.

loadConfig

Arguments

:: FilePath
Caminho do ficheiro de configuração.

-> IO FontConfig

Carrega o ficheiro de configuração de um tipo de letra.

loadChars

Arguments

:: FilePath
Caminho da diretoria.

-> [Color]
Cores suportadas.

-> IO (Map Char FontChar)

Carrega os carateres de um tipo de letra, a partir de uma diretoria adequada.

loadChar

Arguments

:: [Color]
Cores suportadas.

-> FilePath
Caminho do ficheiro.

-> IO (Char, FontChar)

Carrega um caratere de um tipo de letra, a partir de um ficheiro de imagem do tipo BMP.

Desenho de texto

Medição

size

Arguments

:: Font
Tipo de letra.

-> Text
Texto a medir.

-> Vector
Dimensão do texto ao ser desenhado.

Calcula as dimensões que uma secção de texto tomará ao ser desenhada.

sizeString

Arguments

:: Font
Tipo de letra.

-> String
Texto a medir.

-> Vector
Dimensão do texto ao ser desenhado.

Calcula as dimensões que uma String de texto tomaria ao ser desenhada no tamanho nativo do tipo de letra.

widthLine

Arguments

:: Font
Tipo de letra.

-> String
Linha de texto a medir.

-> Float
Largura da linha texto ao ser desenhada.

Calcula a largura que uma String de texto sem newlines tomaria ao ser desenhada no tamanho nativo do tipo de letra.

Transformação em Picture

draw

Arguments

:: Vector
Posição principal do texto. O seu significado varia consoante o alinhamento do texto.

-> Font
Tipo de letra.

-> Text
Texto a ser desenhado.

-> Picture
Picture resultante.

Desenha uma secção de texto.

drawString

Arguments

:: Alignment
Alinhamento do texto.

-> Vector
Posição principal do texto.

-> Font
Tipo de letra.

-> String
Texto a ser desenhado.

-> Float
Escala a aplicar ao texto.

-> Color
Cor do texto.

-> Picture

Desenha uma String.

drawLineLeft :: Font -> Color -> Float -> String -> [Picture]

Desenha uma linha de texto alinhada à esquerda.

drawLineCenter :: Font -> Color -> Float -> String -> [Picture]

Desenha uma linha de texto alinhada à centrada horizontalmente.

drawLineRight :: Font -> Color -> Float -> String -> [Picture]

Desenha uma linha de texto alinhada à direita.

drawLine

Arguments

:: Font
Tipo de letra.

-> Color
Cor do texto.

-> Vector
Canto superior esquerdo da linha.

-> String
Texto a desenhar.

-> [Picture]
Imagens de cada caratere.

Desenha uma linha de texto.