Sokoban.UI.UIHelper

Description

Este módulo contém tipos e funções úteis a vários fins relacionados com a interface do utilizador (UI).

Background

data BackgroundStyle

Define o estilo de um background.

Constructors

Light
Background claro.

Dark
Background escuro.

Darker
Background ainda mais escuro.

createBackground

Arguments

:: Assets
Assets.

-> Rect
Região onde a imagem será desenhada.

-> BackgroundStyle
Estilo do background.

-> Picture
Imagem resultante.

Cria uma imagem de fundo que pode ser utilizada em várias situações.