Tarefa2

Description

O objetivo desta tarefa é implementar um sistema de reação a comandos, que dado um número de um jogador e um movimento, retorne o estado do mapa após

esse moviment ter sido efetuado ou o mesmo estado do jogo, no caso de a jogada não ter sido possível.

Documentation

move :: [String] -> Int -> Char -> [String]

A função move recebe um estado de jogo, um Int e um Char, que indicam o jogador e ajogada pretendida, respetivamente. Ao prcorrer a lista de

Strings correspondente ao estado do jogo, a função verifica qual string é a correspondente ao jogador no argumento, para depois utilizar essa

informação nas suas funções auxiliares. S uma dada jgada é possível, então o mapa é reescrito.

tarefa2 :: IO ()

A função main recebe dois como input os argumentos relativos ao jogador e à jogada, chamando a função move, aqui definida, com esses mesmos argumentos.

possibleM :: [String] -> Char -> Bool

A função possibleM verifica se uma dada jogada, para uma dada lista de Strings, em que a última String é a correspondente ao jogador que

quer efetuar a jogada, é possível.

bombaSpot :: [String] -> String -> Bool

Esta função averigua se há uma bomba nas cordenadas em que o jogador se encontra, dado que não pode haver mais de uma bomba por célula.

numeroBombasP :: [String] -> String -> Int

A função numerobombasP verifica o número de bombas que um dado jogador já tem colocadas num mapa.

replaceRows :: [String] -> Char -> [String]

A função replaceRows é o "coração" da nossa solução para esta tarefa, dado que é ela que reescreve o mapa quandouma jogada é feita, ao eliminar

Strings de powerups que tenham sido apanhados na jogada, adicionando powerups a jogadores, reescrevedo as coordenadas dos mesmos, e também adicionando

bombas, no caso de ser essa a jogada.

pOrden :: String -> Int

Esta pequena função traduz a função pOrdenSt num Int, de forma a podermos trabalhar com esta informação em índices, por exemplo.

pAbcis :: String -> Int

O papel desempenhado pela pAbcis é análogo ao da pOrden, mas diz respeito à abcissa de um dado objeto (e.g. jogador, bomba, power up..)

pAbcisSt :: String -> String

Esta função obtem a parte da string correspondente a um jogador, ou outro objeto com coordenadas, em que se encontra a informação sobre a abcissa

das coordenadas desse mesmo objeto. Como esta informação começa sempre no 3º caracter de cada string, apenas temos de considerar o quão grande é o

número relativo a esta mesma coordenada.

pOrdenSt :: String -> String

Apesar de esta função ter exaamente mesmo papel que a pAbcisSt, a ordenada dos objetos vem em segundo lugar, e, por isso, os caracteres da string

que correspondem a esta informação também dependem da quantidade de caracteres utilizados para descrever a abcissa.

right :: String -> String

Esta função reescreve uma string de um jogador que se mova para a direita, somando 1 à sua abcissa. Funcionando d forma semelhante, definimos também as

funções left, up e down

>right "0 1 1 ++"
"0 2 1 ++"

>down "0 2 1 ++"
"0 2 2 ++"

left :: String -> String

up :: String -> String

down :: String -> String

bomb :: [String] -> String -> [String]

A função bomb e responsável por colocar uma bomba no mapa, recorrendo ao próprio mapa e à string relativa ao jogador que a colocou.

bombSt :: String -> String

Esta função é responsável pela construção da string com a informação de uma nova bomba, a introduzir na lista de Strings com a informação do jogo.

>bombSt "0 1 1 !"
"* 1 1 1 2 10"

numeroVezes :: Char -> String -> Int

Esta função indica quantas vezes um dado char ocorre numa string, o que é bastante útil, por exemplo no casode querermos saber quandos power ups

de um certo tipo um jogador tem.

removePowers :: [String] -> String -> [String]

A função remove powers averigua se há power ups em coordenadas coincidentes com as do jogador (após se mover), removendo as strings

relativas a esse mesmo powrup, dado que foi apanhado pelo jogador.

isPower :: [String] -> String -> Char

Esta função percorre a lista de Strings e, a partir das coordenadas da posição para onde o jogador se vai mover, verifica se há algum powerup nessa localização

para o jogador apanhar.

addPower :: Char -> String -> String

A função addPower adiciona a uma String de um jogador um dado powerup.

startPowers :: String -> Int

A startPowers, a partir de uma String (relativa a um jogador) define o número de caracteres que vêm antes do começo da informação que diz respeito aos powerups.