Tarefa6_li1g156

Description

O objetivo desta tarefa é o desenvolvimento de um bot que jogasse Bomberman sem qualquer apoio humano.

Documentation

type Mapa = [String]

type Cords

Arguments

= (Abcissa, Ordenada)
(Abcissa,Ordenada)

type State

Arguments

= (Mapa, [Cords], [Cords], [Powers], [Bombs], [Players])
(Mapa, [lista de espaços vazios, por onde se pode andar], [lista de tijolos], [lista de powerups], [lista de bombas], [lista de players])

type Bombs

Arguments

= (Int, [Cords])
(Instantes que faltam para explodir, [lista de espaços que vão arder])

type Players

Arguments

= (Cords, Int, Int)
(Coordenadas, numero de bombas que pode pôr, range das bombas).

type Powers

Arguments

= (Char, Cords)
(Tipo de powerup (1=bombs,2=flames), coordenadas).

type Caminho

Arguments

= String
Caminho desde um espaço do mapa a outro, sob a forma de uma lista de instruções (chars).

type Abcissa

Arguments

= Int
Abcissa de um espaço no mapa.

type Ordenada

Arguments

= Int
Ordenada de um espaço no mapa.

type Range

Arguments

= Int
Raio de uma bomba colocada no mapa.

type Time

Arguments

= Int
Instantes de tempo para o jogo terminar.

type Dim

Arguments

= Int
Dimensão do mapa.

type Jogador

Arguments

= Int
Número que representa um dado jogador.

type Linha

Arguments

= String
Uma linha da representação doe stado do jogo no tipo Mapa

data RPath a

Novo tipo de dados para tornar possível a criação de um variado conjunto de movimentos "óptimos" para o bot.

Constructors

C a [RPath a]

bot :: Mapa -> Time -> Jogador -> Maybe Char

botState1 :: State -> Jogador -> Maybe Char

Função que determina a ação do bot quando o mapa ainda não tem um tamanho demasiado reduzido.

botState2 :: State -> Time -> Jogador -> Maybe Char

Função que determina a ação do bot quando o mapa se encontra bastante reduzido devido à espiral.

safe :: State -> Jogador -> Char -> Maybe Char

Função que verifica se uma jogada leva um jogador a estar em risco de morrer para uma bomba.

timers :: Cords -> State -> Int

Função que averigua quanto tempo falta para a próxima bomba que explode num determinado local explodir.

auxtimers :: [Bombs] -> Cords -> [Int]

Auxiliar da função timers

saferSpots :: [Cords] -> State -> Time -> Cords -> [Cords]

Função que tem em conta quais espaços serão engolidos primeiro pela espiral.

returnIndex :: Eq a => [a] -> a -> Int

Função que devolve o índice de um elemento numa lista.

worseTimerWay :: State -> Caminho -> Cords -> Bool

Função que testa se um dado caminho leva a um maior risco para jogador.

unsafeSpiral :: Time -> Dim -> [Cords]

Função que verifica os espaços a ser esmagados pela espiral nos próximos instantes.

lastSquare :: State -> Cords

Função que determina o último espaço do mapa e que um jogador pode estar antes de ser exmagado pela espiral.

immaDie :: State -> Cords -> Bool

Função que testa se uma dada posição está em risco de ser abrangida por uma explosão.

immaDie1 :: Time -> State -> Cords -> Bool

Função que testa se uma dada posição está em risco de ser abrangida por uma explosão, quando o tempo está a acabar.

safeSpots :: State -> [Cords] -> [Cords]

Função que encontra numa lista de coordenadas um par de coordenadas que corresponda a um sítio que não esteja em risco de explosão.

chooseSpot :: State -> [Cords] -> Cords -> Cords

Função que seleciona o espaço seguro mais próximo.

chooseSpot1 :: Time -> State -> [Cords] -> Cords -> Cords

Função que seleciona o espaço seguro mais próximo.

findSafeSpot :: Int -> State -> Cords -> Cords

Função que encontra as coordenadas seguras mais próximas de umas certas coordenadas.

findSafeSpot1 :: Time -> Int -> State -> Cords -> Cords

Função que encontra as coordenadas seguras mais próximas de umas certas coordenadas, quando o tempo está a acabar.

runAway :: Int -> State -> Cords -> Caminho

Função que elabora um caminho para fugir às zonas prestes a ser abrangidas por explosões.

runAway1 :: Time -> Int -> State -> Cords -> Caminho

Função que elabora um caminho para fugir às zonas prestes a ser abrangidas por explosões, quando o tempo está a acabar.

findPoint :: [Cords] -> Cords -> Cords -> Cords

Encontra o ponto mais próximo de um destino numa certa rede.

dist :: Cords -> Cords -> Float

Distancia entre 2 espaços, dadas as suas coordenadas.

findWay :: Int -> State -> Cords -> Cords -> Caminho

Função que encontra um caminho entre umas dadas coordenadas e outras coordenadas.

mkDirections :: [Cords] -> Caminho

Função que converte uma lista de coordenadas numa lista de movimentos do jogador.

pTreeToCords :: RPath Cords -> [Cords]

Função que converte uma árvore de caminhos (com apenas um ramo) numa lista de coordenadas.

mkPathTree :: Int -> Cords -> Cords -> [Cords] -> RPath Cords

Função que cria uma árvore de caminhos numa rede que partem de um nodo.

choosePath :: RPath Cords -> RPath Cords

Fução que escolhe o caminho mais curto.

pathHeight :: RPath Cords -> Int

Função que calcula o numero de nodos de um caminho.

prunePaths :: Cords -> RPath Cords -> RPath Cords

Função que retira de uma lista de caminhos os becos sem saída.

prunePaths1 :: Cords -> RPath Cords -> RPath Cords

Função que retira de uma lista de caminhos os becos sem saída.

concatPTree :: RPath a -> RPath a -> RPath a

Função que junta duas árvores de caminhos.

exc :: Eq a => [a] -> [a] -> [a]

Função que excui os elementos de uma lista de outra.

neighbours :: Cords -> [Cords] -> [Cords]

network :: Int -> [Cords] -> Cords -> [Cords]

Função que cria uma rede de espaços ligados entre si a partir de uma lista de epaços. Note-se que o Int que esta função leva como argumento serve para limitar os cálculos a efetuar, dado que estes crescem a um ritmo exponencial, mediante a quantidade de nodos de uma rede.

network1 :: Int -> Cords -> [Cords] -> [Cords] -> [Cords]

Função auxiliar da network

isNode :: Cords -> [Cords] -> Bool

Função que testa se um par de coordenadas está numa rede de espaços ligados entre si.

testWay :: Mapa -> Jogador -> Cords -> Caminho -> Bool

Função que testa se uma string com instruções para um caminho a ser seguido por um jogador o levam a determnadas coordenadas.

walkWay :: Mapa -> Jogador -> Caminho -> Mapa

Função que testa um caminho utilizando a função move da Tarefa2.

mkState :: Mapa -> State

Função que a partir de uma lista de strings cria uma estrutura de dados representativa do estado do jogo mais conveniente para trabalhar.

takeCordsChar :: Char -> Ordenada -> Mapa -> [Cords]

Função que retira do mapa elementos representados pelo Char x em forma de túpulo.

takeCordsCharRow :: Char -> Ordenada -> Abcissa -> Linha -> [Cords]

Função que retira de cada linha do mapa um dado elemento representado por dado Char, em forma de túpulo.

retiraPowers :: Mapa -> [Powers]

Função que a partir do estado do jogo obtém informação dos powerups.

tuplePowers :: Linha -> Powers

Função que converte num túpulo uma string com informação de dum dado powerup.

takeBombs :: Mapa -> Mapa -> [Bombs]

Função que retira do mapa informação acerca das bombas.

cordFlames2 :: Mapa -> Linha -> [Cords]

Função que cria uma lista de coordenadas em que são propagadas as chamas.

cordFlamesUp2 :: Mapa -> Range -> Abcissa -> Ordenada -> [Cords]

cordFlamesDown2 :: Mapa -> Range -> Abcissa -> Ordenada -> [Cords]

cordFlamesLeft2 :: Mapa -> Range -> Abcissa -> Ordenada -> [Cords]

cordFlamesRight2 :: Mapa -> Range -> Abcissa -> Ordenada -> [Cords]

takePlayers :: Mapa -> Mapa -> [Players]

Função que obtem informação relativa aos jogadores em jogo.

cordsPlayer :: Mapa -> Jogador -> Cords

Função que a partir de um estado e de um numero de jogador obtém um túpulo com as coordenadas do jogador.

howMany :: Char -> Linha -> Int

Função que conta o número de vezes que um certo caracter ocorre numa string, com utilididade, por exemplo, no caso de saber quantos powerups de determinado tipo tem um jogador.

rmDups :: Eq a => [a] -> [a]

Função que remove elementos repetidos numa lista.