TarefaA

Description

Este módulo contém uma resolução correspondente à Tarefa A da primeira fase do projeto de Laboratórios de Informática I.

O objetivo é o de verificar se o input fornecido através do stdin corresponde ao formato especificado pelo enunciado do projeto, devolvendo através do stdout uma mensagem de sucesso ou o número da linha onde for encontrada a primeira divergência.

Entrada

tarefaA :: [String] -> [String]

Função inicial da tarefa. Recebe o input do programa e devolve o output correspondente.

Tipos Auxiliares

Tipos genéricos utilizados pelo programa.

type Point = (Int, Int)

Representa um ponto no plano.

type Size = (Int, Int)

Representa uma dimensão no plano.

Processamento do Input

Tipos e funções destinados ao processamento e manipulação do input do programa.

data Line

Representa uma linha de input numerada.

Constructors

Line

Fields

lineNum :: LineNumber

O número da linha.

lineText :: String

O conteúdo da linha.

type BoardLine = Line

Sinónimo de Line, utilizado para distinguir linhas correspondentes à secção do tabuleiro.

type CoordsLine = Line

Sinónimo de Line, utilizado para distinguir linhas correspondentes à secção das coordenadas.

type LineNumber = Int

Representa o número de uma linha de input.

toLineList :: [String] -> [Line]

Converte uma lista de Strings numa lista de linhas numeradas.

Exemplos de utilização:

>>> toLineList []
[]

>>> toLineList ["ABC", "123"]
[ Line {lineNum = 1, lineText = "ABC"},Line {lineNum = 2, lineText = "123"} ]

>>> toLineList ["#####", "#  .#", "#####"]
[ Line {lineNum = 1, lineText = "#####"},
Line {lineNum = 2, lineText = "#  .#"},
Line {lineNum = 3, lineText = "#####"} ]

lineIsNotEmpty :: Line -> Bool

Verifica se um valor do tipo Line representa uma linha não vazia.

Exemplos de utilização:

>>> lineIsNotEmpty (Line 42 "")
False

>>> lineIsNotEmpty (Line 42 "Ola!")
True

splitInput :: [Line] -> ([BoardLine], [CoordsLine])

Divide uma lista de linhas de input em duas secções.

Se as linhas representarem corretamente um tabuleiro de jogo e as coordenadas do boneco e das caixas, a primeira secção corresponderá ao tabuleiro e a segunda às coordenadas.

Tabuleiro

Tipos e funções específicos ao processamento de tabuleiros de jogo.

data Board

Representa um tabuleiro.

Constructors

Board

Fields

boardSize :: Size

O tamanho do tabuleiro.

boardCells :: [[Cell]]

As células do tabuleiro.

type Cell = Char

Representa uma célula de um tabuleiro.

Pode tomar os seguintes valores:

'#' - a célula é uma parede;

' ' - a célula é vazia;

'.' - a célula é um local de arrumação.

parseBoard :: [Line] -> Board

Constrói um valor do tipo Board a partir de uma representação textual válida.

Exemplos de utilização:

>>> parseBoard [Line 1 "#####", Line 2 "#  .#", Line 3 "#####"]
Board { boardSize = (5,3), boardCells = ["#####","#  .#","#####"] }

getCell

Arguments

:: Point
Coordenadas da célula.

-> Board
Tabuleiro.

-> Cell
Célula nas coordenadas especificadas.

Devolve a célula de um tabuleiro com as coordenadas especificadas.

Exemplos de utilização:

>>> getCell (3,1) (parseBoard [Line 1 "#####",Line 2 "#  .#",Line 3 "#####"])
'.'

>>> getCell (0,1) (parseBoard [Line 1 "#####",Line 2 "#  .#",Line 3 "#####"])
'#'

numTargets :: Board -> Int

Calcula o número de locais de arrumação presentes num tabuleiro, ou seja, o número de células de valor '.' existentes no tabuleiro.

Exemplos de utilização:

>>> numTargets (parseBoard [Line 1 "#####", Line 2 "#   #", Line 3 "#####"])
0

>>> numTargets (parseBoard [Line 1 "#####", Line 2 "#  .#", Line 3 "#####"])
1

>>> numTargets (parseBoard [Line 1 "#####", Line 2 "#. .#", Line 3 "#####"])
2

Coordenadas

isValidCoordFormat :: Line -> Bool

Verifica se uma linha de input representa corretamente um par de coordenadas.

Exemplos de utilização:

>>> isValidCoordFormat (Line 42 "")
False

>>> isValidCoordFormat (Line 42 "5 b")
False

>>> isValidCoordFormat (Line 42 "1 3 7")
False

>>> isValidCoordFormat (Line 42 "1 3")
True

Testes

data TestResult

Representa o resultado de um teste.

Constructors

Ok
O teste passou.

Error LineNumber
O teste falhou na linha de número especificado.

firstError :: [TestResult] -> TestResult

Devolve o primeiro elemento com a forma Error n de uma lista do tipo [TestResult].

Se a lista não possuir elementos com essa forma, é devolvido o valor Ok.

Exemplos de utilização:

>>> firstError []
Ok

>>> firstError [Ok, Ok]
Ok

>>> firstError [Error 7, Error 1]
Error 7

>>> firstError [Error 4, Error 2, Ok, Error 9]
Error 4

lowestError :: [TestResult] -> TestResult

Devolve o elemento com a forma Error n de menor n, de uma lista do tipo [TestResult].

Se a lista não possuir elementos com essa forma, é devolvido o valor Ok.

Exemplos de utilização:

>>> lowestError []
Ok

>>> lowestError [Ok, Ok]
Ok

>>> lowestError [Error 7, Error 1]
Error 1

>>> lowestError [Error 4, Error 2, Ok, Error 9]
Error 2

testResultToString :: TestResult -> String

Converte um valor do tipo TestResult numa representação textual equivalente.

Exemplos de utilização:

>>> testResultToString Ok
"OK"

>>> testResultToString (Error 42)
"42"

testInput

Arguments

:: [Line]
Input do programa.

-> TestResult
Resultado do teste.

Testa se o input do programa segue o formato especificado pelo enunciado do projeto.

Se o input seguir o formato em questão, é devolvido o valor Ok. Em caso contrário, é devolvido o valor Error n, onde n corresponde ao número da linha onde for encontrada a primeira divergência.

Tabuleiro

testBoard

Arguments

:: [BoardLine]
Linhas de input a testar.

-> TestResult
Resultado do teste.

Testa se a secção do input do programa correspondente ao tabuleiro segue o formato especificado pelo enunciado do projeto.

Se o input seguir o formato em questão, é devolvido o valor Ok. Em caso contrário, é devolvido o valor Error n, onde n corresponde ao número da linha onde for encontrada a primeira divergência.

testBoardChars

Arguments

:: [BoardLine]
Linhas de input a testar.

-> TestResult
Resultado do teste.

Testa se todas as linhas de uma lista de linhas de input apenas contêm carateres válidos.

Os únicos carateres válidos são os seguintes: '#', ' ', '.'.

Se não forem encontrados carateres inválidos, é devolvido o valor Ok. Em caso contrário, é devolvido o valor Error n, onde n corresponde ao número da primeira linha que contém um carater inválido.

Exemplos de utilização:

>>> testBoardChars [Line 1 "#####", Line 2 "#  .#", Line 3 "#####"]
Ok

>>> testBoardChars [Line 1 "#####", Line 2 "# X.#", Line 3 "#####"]
Error 2

>>> testBoardChars [Line 1 "#####", Line 2 "# X.#", Line 3 "#Ola#"]
Error 2

testBoardShape

Arguments

:: [BoardLine]
Linhas de input a testar.

-> TestResult
Resultado do teste.

Testa se todas as linhas de uma lista de linhas de input são não vazias e possuem o mesmo comprimento.

Se tal se verificar, é devolvido o valor Ok. Em caso contrário, é devolvido o valor Error n, onde n corresponde ao número da primeira linha vazia ou de comprimento diferente à anterior.

Exemplos de utilização:

>>> testBoardShape [Line 1 "#####", Line 2 "#  .#", Line 3 "#####"]
Ok

>>> testBoardShape [Line 1 "#####", Line 2 "#  .#", Line 3 ""]
Error 3

>>> testBoardShape [Line 1 "######", Line 2 "#  .#", Line 3 "######"]
Error 2

testBoardBounds

Arguments

:: [BoardLine]
Linhas de input a testar.

-> TestResult
Resultado do teste.

Testa se a primeira e última linhas de uma lista de linhas de input apenas contêm o carater '#' e as restantes são prefixadas e sufixadas pelo carater '#'.

Se tal se verificar, é devolvido o valor Ok. Em caso contrário, é devolvido o valor Error n, onde n corresponde ao número da primeira linha que não siga o formato esperado.

Exemplos de utilização:

>>> testBoardBounds [Line 1 "#####", Line 2 "#  .#", Line 3 "#####"]
Ok

>>> testBoardBounds [Line 1 "#####", Line 2 "#  . ", Line 3 "#####"]
Error 2

>>> testBoardBounds [Line 1 "#####", Line 2 "#  .#", Line 3 "#   #"]
Error 3

Coordenadas

testCoords

Arguments

:: Board
Tabuleiro.

-> [CoordsLine]
Linhas de input a testar.

-> TestResult
Resultado do teste.

Testa se a secção do input do programa correspondente às coordenadas segue o formato especificado pelo enunciado do projeto, tendo em conta um tabuleiro.

Se o input seguir o formato em questão, é devolvido o valor Ok. Em caso contrário, é devolvido o valor Error n, onde n corresponde ao número da linha onde for encontrada a primeira divergência.

testCoordsEmptyLines

Arguments

:: [CoordsLine]
Linhas de input a testar.

-> TestResult
Resultado do teste.

Testa se não existem linhas não vazias após linhas vazias, numa lista de linhas de input.

Se após a primeira a linha vazia apenas existirem linhas vazias, é devolvido o valor Ok. Em caso contrário, é devolvido o valor Error n, onde n corresponde ao número da primeira linha não vazia após a primeira linha vazia.

Exemplos de utilização:

>>> testCoordsEmptyLines []
Ok

>>> testCoordsEmptyLines [Line 13 "A", Line 14 "", Line 15 ""]
Ok

>>> testCoordsEmptyLines [Line 13 "A", Line 14 "", Line 15 "C"]
Error 15

testCoordsCount

Arguments

:: Board
Tabuleiro.

-> [CoordsLine]
Linhas de input a testar.

-> TestResult
Resultado do teste.

Testa se o número de coordenadas descritas por uma lista de linhas é válido, tendo em conta o número de locais de arrumação num tabuleiro.

As linhas devem conter as coordenadas do boneco e as coordenadas de cada caixa. Devem existir tantas caixas como locais de arrumação. Assim, o número de linhas deve ser dado pela expressão:

1 + nº de locais de arrumação no tabuleiro

Se o número de linhas for igual ao número esperado, é devolvido o valor Ok.

Se o número de linhas for inferior ao número esperado, é devolvido o valor Error n, onde n corresponde ao número da linha que deveria existir após a última linha na lista.

Se o número de linhas for superior ao número esperado, é devolvido o valor Error n, onde n corresponde ao número da primeira linha em excesso.

testCoordsLines

Arguments

:: Board
Tabuleiro.

-> [CoordsLine]
Linhas de input a testar.

-> TestResult
Resultado do teste.

Testa se todas as linhas de uma lista de linhas de input representam corretamente um par de coordenadas válido.

Para um par de coordenadas ser válido, este deve corresponder a uma posição dentro do tabuleiro especificado e a uma célula que não seja uma parede. Não deve existir também mais de uma linha a representar o mesmo ponto.

Se todas as linhas forem válidas, é devolvido o valor Ok. Em caso contrário, é devolvido o valor Error n, onde n corresponde ao número da primeira linha não válida.

testCoordsLine

Arguments

:: Board
Tabuleiro.

-> CoordsLine
Linha de input a testar.

-> [Point]
Lista de pontos inválidos.

-> Either [Point] LineNumber
Resultado do teste.

Testa se uma linha de input representa corretamente um par de coordenadas válido.

Para o par de coordenadas ser válido, este deve corresponder a uma posição dentro do tabuleiro especificado e a uma célula que não seja uma parede. Não se deve encontrar também na lista de pontos especificada.

Se a linha passar o teste, é devolvido o valor Left pts, onde pts corresponde à lista de pontos passada como argumento, aumentada com o par de coordenadas que a linha representa.

Se a linha falhar o teste, é devolvido o valor Right n, onde n corresponde ao número da linha.