TarefaB

Description

Este módulo contém uma resolução correspondente à Tarefa B da primeira fase do projeto de Laboratórios de Informática I.

O objetivo é o de apresentar no stdout uma visualização do tabuleiro de jogo que inclua as caixas e o boneco, a partir do input fornecido através do stdin.

Entrada

tarefaB :: [String] -> [String]

Função inicial da tarefa. Recebe o input do programa e devolve o output correspondente.

Tipos e Funções Auxiliares

Tipos e funções genéricos utilizados pelo programa.

type Point = (Int, Int)

Representa um ponto no plano.

type Size = (Int, Int)

Representa uma dimensão no plano.

replaceElemAt

Arguments

:: Int
Índice do elemento a substituir.

-> a
Valor pelo qual substituir o elemento.

-> [a]
Lista original.

-> [a]
Lista resultante.

Substitui um elemento de uma lista pelo valor especificado.

Exemplos de utilização:

>>> replaceElemAt 1 'E' ['a','b','c','d']
['a','E','c','d']

>>> replaceElemAt 3 9 [1,2,3,4]
[1,2,3,9]

Processamento do Input

Tipos e funções destinados ao processamento e manipulação do input do programa.

parseInput :: [String] -> (Board, Coords)

Contrói um tuplo com valores dos tipos Board e Coords a partir de uma representação textual adequada.

Exemplos de utilização:

>>> parseInput ["#####", "#  .#", "#####", "1 1", "2 1", "3 1"]
( Board { boardSize = (5,3), boardCells = ["#####","#  .#","#####"] },
Coords { coordsPlayer = (1,1), coordsBoxes = [(2,1),(3,1)] } )

Tabuleiro

Tipos e funções específicos ao processamento de tabuleiros de jogo.

data Board

Representa um tabuleiro.

Constructors

Board

Fields

boardSize :: Size

O tamanho do tabuleiro.

boardCells :: [[Cell]]

As células do tabuleiro.

type Cell = Char

Representa uma célula de um tabuleiro.

parseBoard :: [String] -> Board

Constrói um valor do tipo Board a partir de uma representação textual adequada.

Exemplos de utilização:

>>> parseBoard ["#####", "#  .#", "#####"]
Board { boardSize = (5,3), boardCells = ["#####","#  .#","#####"] }

getCell

Arguments

:: Point
Coordenadas da célula.

-> Board
Tabuleiro.

-> Cell
Célula nas coordenadas especificadas.

Devolve a célula de um tabuleiro com as coordenadas especificadas.

Exemplos de utilização:

>>> getCell (3,1) (parseBoard ["#####", "#  .#", "#####"])
'.'

>>> getCell (0,1) (parseBoard ["#####", "#  .#", "#####"])
'#'

setCell

Arguments

:: Cell
Valor pelo qual substituir a célula.

-> Point
Coordenadas da célula.

-> Board
Tabuleiro original.

-> Board
Tabuleiro resultante.

Substitui uma célula de um tabuleiro pelo valor especificado e devolve o tabuleiro resultante.

Exemplos de utilização:

>>> setCell ' ' (3,1) (parseBoard ["#####", "#  .#", "#####"])
Board { boardSize = (5,3), boardCells = ["#####","#   #","#####"] }

>>> setCell '#' (2,1) (parseBoard ["#####", "#  .#", "#####"])
Board { boardSize = (5,3), boardCells = ["#####","# #.#","#####"] }

cellIsWall

Arguments

:: Board
Tabuleiro.

-> Point
Coordenadas da célula.

-> Bool
True se a célula em questão for considerada uma parede. Em caso contrário, False.

Verifica se uma célula de um tabuleiro é uma parede.

Se as coordenadas especificadas não corresponderem a uma posição válida dentro do tabuleiro, considera-se que o resultado é verdadeiro.

Exemplos de utilização:

>>> cellIsWall (parseBoard ["#####", "#  .#", "#####"]) (1,1)
False

>>> cellIsWall (parseBoard ["#####", "#  .#", "#####"]) (1,2)
True

>>> cellIsWall (parseBoard ["#####", "#  .#", "#####"]) (-1,-3)
True

removeUnnecessaryWalls

Arguments

:: Board
Tabuleiro original.

-> Board
Tabuleiro resultante.

Substitui por uma célula vazia as células de um tabuleiro que sejam paredes e estejam rodeadas por paredes.

Exemplos de utilização:

>>> removeUnnecessaryWalls (parseBoard ["#####", "## .#", "#  ##", "#####"])
Board { boardSize = (5,4), boardCells = [" ####","## .#","#  ##","#### "] }

removeIfIsUnnecessaryWall

Arguments

:: Board
Tabuleiro para a verificação.

-> Point
Coordenadas da célula.

-> Board
Tabuleiro a ser modificado.

-> Board
Tabuleiro resultante.

Substitui uma célula por uma célula vazia se esta for uma parede e estiver rodeada por paredes.

A verificação do valor da célula e das células à sua volta é feita sobre um tabuleiro, mas a alteração é feita a outro tabuleiro. Assim, a função recebe dois tabuleiros.

Coordenadas

Tipos e funções relacionados com o processamento de coordenadas.

data Coords

Representa as coordenadas (posições no tabuleiro) do boneco e de todas as caixas.

Constructors

Coords

Fields

coordsPlayer :: Point

As coordenadas do boneco.

coordsBoxes :: [Point]

As coordenadas de todas as caixas.

parseCoords :: [String] -> Coords

Constrói um valor do tipo Coords a partir de uma representação textual adequada.

Exemplos de utilização:

>>> parseCoords ["11 2", "5 8", "7 7", "5 6"]
Coords { coordsPlayer = (11,2), coordsBoxes = [(5,8),(7,7),(5,6)] }

applyCoordsToBoard

Arguments

:: Coords
Coordenadas do boneco e das caixas.

-> Board
Tabuleiro original.

-> Board
Tabuleiro resultante.

Modifica um tabuleiro para que nele fiquem representados o boneco e as caixas nas células em que estes se encontrem.

A célula na qual o boneco se encontra (a qual se assume estar vazia) passa a ter o valor 'o'.

As células nas quais se encontram caixas passam a ter o valor:

'H' se a célula estiver vazia;

'I' se a célula for um local de arrumação.

applyPlayerToBoard

Arguments

:: Point
Coordenadas do boneco.

-> Board
Tabuleiro original.

-> Board
Tabuleiro resultante.

Modifica um tabuleiro para que nele fique representado o boneco na célula em que este se encontre.

A célula na qual o boneco se encontra (a qual se assume estar vazia) passa a ter o valor 'o'.

applyBoxesToBoard

Arguments

:: [Point]
Coordenadas das caixas.

-> Board
Tabuleiro original.

-> Board
Tabuleiro resultante.

Modifica um tabuleiro para que nele fiquem representadas as caixas nas células em que estas se encontrem.

As células nas quais se encontram caixas passam a ter o valor:

'H' se a célula estiver vazia;

'I' se a célula for um local de arrumação.

applyBoxToBoard

Arguments

:: Point
Coordenadas da caixa.

-> Board
Tabuleiro original.

-> Board
Tabuleiro resultante.

Modifica um tabuleiro para que nele fique representada uma caixa na célula em que esta se encontre.

A célula na qual se encontra a caixa passa a ter o valor:

'H' se a célula estiver vazia;

'I' se a célula for um local de arrumação.

makePrettyBoard

Arguments

:: Board
Tabuleiro original.

-> Coords
Coordenadas do boneco e das caixas.

-> Board
Tabuleiro resultante.

Modifica um tabuleiro para que nele fiquem representados o boneco e as caixas, substituindo também por célula vazias as paredes desnecessárias.

A célula na qual o boneco se encontra (a qual se assume estar vazia) passa a ter o valor 'o'.

As células nas quais se encontram caixas passam a ter o valor:

'H' se a célula estiver vazia;

'I' se a célula for um local de arrumação.

São consideradas desnecessárias as células que sejam paredes e estejam rodeadas por paredes.