TarefaC

Description

Este módulo contém uma resolução correspondente à Tarefa C da primeira fase do projeto de Laboratórios de Informática I.

O objetivo é o de apresentar no stdout a posição do boneco após a execução (ou tentativa de execução) de um comando, a partir do input fornecido através do stdin.

Entrada

tarefaC :: [String] -> [String]

Função inicial da tarefa. Recebe o input do programa e devolve o output correspondente.

Tipos e Funções Auxiliares

Tipos e funções genéricos utilizados pelo programa.

type Vector = (Int, Int)

Representa um vetor bidimensional.

Pode ser utilizado para representar pontos, translações, direções, dimensões, etc.

(|+|) :: Vector -> Vector -> Vector

Adiciona dois valores do tipo Vector componente por componente, segundo a equação:

(x1, y1) |+| (x2, y2) = (x1+x2, y1+y2)

Exemplos de utilização:

>>> (1,4) |+| (2,5)
(3,9)

vectorToString :: Vector -> String

Converte um valor do tipo Vector numa representação textual equivalente.

Exemplos de utilização:

>>> vectorToString (1,2)
"1 2"

>>> vectorToString (-3,-4)
"-3 -4"

replaceElemAt

Arguments

:: Int
Índice do elemento a substituir.

-> a
Valor pelo qual substituir o elemento.

-> [a]
Lista original.

-> [a]
Lista resultante.

Substitui um elemento de uma lista pelo valor especificado.

Exemplos de utilização:

>>> replaceElemAt 1 'E' ['a','b','c','d']
['a','E','c','d']

>>> replaceElemAt 3 9 [1,2,3,4]
[1,2,3,9]

Processamento do Input

Tipos e funções destinados ao processamento e manipulação do input do programa.

parseInput :: [String] -> (Board, Coords, Command)

Constrói um tuplo com valores dos tipos Board, Coords e Command a partir de uma representação textual adequada.

Exemplos de utilização:

>>> parseInput ["#####", "#  .#", "#####", "1 1", "2 1", "3 1", "R"]
( Board { boardSize = (5,3), boardCells = ["#####","#  .#","#####"] },
Coords { coordsPlayer = (1,1), coordsBoxes = [(2,1),(3,1)] },
'R' )

Tabuleiro

Tipos e funções específicos ao processamento de tabuleiros de jogo.

data Board

Representa um tabuleiro.

Constructors

Board

Fields

boardSize :: Vector

O tamanho do tabuleiro.

boardCells :: [[Cell]]

As células do tabuleiro.

type Cell = Char

Representa uma célula de um tabuleiro.

parseBoard :: [String] -> Board

Constrói um valor do tipo Board a partir de uma representação textual adequada.

Exemplos de utilização:

>>> parseBoard ["#####", "#  .#", "#####"]
Board { boardSize = (5,3), boardCells = ["#####","#  .#","#####"] }

getCell

Arguments

:: Vector
Coordenadas da célula.

-> Board
Tabuleiro.

-> Cell
Célula nas coordenadas especificadas.

Devolve a célula de um tabuleiro com as coordenadas especificadas.

Exemplos de utilização:

>>> getCell (3,1) (parseBoard ["#####", "#  .#", "#####"])
'.'

>>> getCell (0,1) (parseBoard ["#####", "#  .#", "#####"])
'#'

Coordenadas

Tipos e funções relacionados com o processamento de coordenadas.

data Coords

Representa as coordenadas (posições no tabuleiro) do boneco e de todas as caixas.

Constructors

Coords

Fields

coordsPlayer :: Vector

As coordenadas do boneco.

coordsBoxes :: [Vector]

As coordenadas de todas as caixas.

parseCoords :: [String] -> Coords

Constrói um valor do tipo Coords a partir de uma representação textual adequada.

Exemplos de utilização:

>>> parseCoords ["11 2", "5 8", "7 7", "5 6"]
Coords { coordsPlayer = (11,2), coordsBoxes = [(5,8),(7,7),(5,6)] }

Comando

Tipos e funções relacionados com o comando que deve ser executado.

type Command = Char

Representa um comando.

Pode tomar os seguintes valores: 'L', 'R', 'U', 'D'.

parseCommand :: String -> Command

Contrói um valor do tipo Command a partir de uma representação textual adequada.

Exemplos de utilização:

>>> parseCommand "L"
'L'

>>> parseCommand "R"
'R'

>>> parseCommand "U"
'U'

>>> parseCommand "D"
'D'

commandDir :: Command -> Vector

Converte um valor do tipo Command num valor do tipo Vector que representa a direção correspondente a esse comando.

Exemplos de utilização:

>>> commandDir 'L'
(-1,0)

>>> commandDir 'R'
(1,0)

>>> commandDir 'U'
(0,1)

>>> commandDir 'D'
(0,-1)

runCommand

Arguments

:: Board
Tabuleiro.

-> Coords
Coordenadas do boneco e das caixas.

-> Command
Comando a executar.

-> Vector
Coordenadas do boneco após a execução do comando.

Executa (ou tenta executar) um comando, dado um certo tabuleiro e coordenadas. Devolve as coordenadas do boneco após a execução do comando.

O boneco apenas se pode movimentar se:

a célula de destino estiver vazia, OU

a célula de destino contiver uma caixa mas a célula imediatamente a seguir estiver vazia.

Se não for possível mover o boneco na direção representada pelo comando, as suas coordenadas permanecem inalteradas.