TarefaD

Description

Este módulo contém uma resolução correspondente à Tarefa D da segunda fase do projeto de Laboratórios de Informática I.

O objetivo é o de ler do stdin o estado de um tabuleiro de jogo e uma sequência de comandos que descrevem o movimento do jogador. Os comandos são depois executados até não haver mais comandos ou o jogo atingir um estado terminal. É apresentado no stdout o número de comandos executados e a indicação de que o jogo foi terminado ou não.

Entrada

tarefaD :: [String] -> [String]

Função inicial da tarefa. Recebe o input do programa e devolve o output correspondente.

Tipos e Funções Auxiliares

Tipos e funções genéricos utilizados pelo programa.

type Vector = (Int, Int)

Representa um vetor bidimensional.

Pode ser utilizado para representar pontos, translações, direções, dimensões, etc.

(|+|) :: Vector -> Vector -> Vector

Adiciona dois valores do tipo Vector componente por componente, segundo a equação:

(x1, y1) |+| (x2, y2) = (x1+x2, y1+y2)

Exemplos de utilização:

>>> (1,4) |+| (2,5)
(3,9)

replaceElemAt

Arguments

:: Int
Índice do elemento a substituir.

-> a
Valor pelo qual substituir o elemento.

-> [a]
Lista original.

-> [a]
Lista resultante.

Substitui um elemento de uma lista pelo valor especificado.

Exemplos de utilização:

>>> replaceElemAt 1 'E' ['a','b','c','d']
['a','E','c','d']

>>> replaceElemAt 3 9 [1,2,3,4]
[1,2,3,9]

Processamento do Input

Tipos e funções destinados ao processamento e manipulação do input do programa.

parseInput :: [String] -> (State, [Command])

Contrói um tuplo com valores dos tipos State e [Command] a partir de uma representação textual adequada.

Estado de Jogo

data State

Armazena o estado de um mapa de jogo, i.e. o tabuleiro, a posição do jogador e o número de comandos já executados.

Constructors

State

Fields

stateBoard :: Board

Tabuleiro de jogo.

statePlayer :: Vector

Posição do jogador.

stateTickCount :: Int

Número de comandos já executados.

stateIsSolved :: State -> Bool

Verifica se um mapa de jogo está num estado terminal.

parseState :: [String] -> State

Constrói um valor do tipo State a partir de uma representação textual adequada.

Exemplos de utilização:

>>> parseState ["######", "#  .  #", "######", "2 1", "5 1"]
State {
    stateBoard     = Board { boardSize = (6,3), boardCells = ["######","#  . H#","######"] },
    statePlayer    = (2,1),
    stateTickCount = 0
    }

showState :: State -> String

Transforma um valor do tipo State numa String que indica se o mapa de jogo está num estado terminal e o número de comandos executados.

Tabuleiro

Tipos e funções específicos ao processamento de tabuleiros de jogo.

data Board

Representa um tabuleiro.

Constructors

Board

Fields

boardSize :: Vector

O tamanho do tabuleiro.

boardCells :: [[Cell]]

As células do tabuleiro.

type Cell = Char

Representa uma célula de um tabuleiro.

parseBoard :: [String] -> Coords -> Board

Contrói um valor do tipo Board a partir de uma representação textual adequada.

getCell

Arguments

:: Vector
Coordenadas da célula.

-> Board
Tabuleiro.

-> Cell
Célula nas coordenadas especificadas.

Devolve a célula de um tabuleiro com as coordenadas especificadas.

setCell

Arguments

:: Cell
Valor pelo qual substituir a célula.

-> Vector
Coordenadas da célula.

-> Board
Tabuleiro original.

-> Board
Tabuleiro resultante.

Substitui uma célula de um tabuleiro pelo valor especificado e devolve o tabuleiro resultante.

cellIsClear :: Cell -> Bool

Verifica se uma célula se encontra vazia.

Exemplos de utilização:

>>> cellIsClear ' '
True

>>> cellIsClear '.'
True

>>> cellIsClear 'H'
False

>>> cellIsClear 'I'
False

>>> cellIsClear '#'
False

cellIsBox :: Cell -> Bool

Verifica se uma célula contém uma caixa.

Exemplos de utilização:

>>> cellIsBox ' '
False

>>> cellIsBox '.'
False

>>> cellIsBox 'H'
True

>>> cellIsBox 'I'
True

>>> cellIsBox '#'
False

boardPutBox :: Vector -> Board -> Board

Coloca uma caixa na célula de um tabuleiro na posição especificada.

Exemplos de utilização:

>>> boardPutBox (1,1) (Board { boardSize = (5,3), boardCells = ["#####","# H.#","#####"] })
Board { boardSize = (5,3), boardCells = ["#####","#HH.#","#####"] }

boardRemoveBox :: Vector -> Board -> Board

Remove uma caixa da célula de um tabuleiro na posição especificada.

Exemplos de utilização:

>>> boardRemoveBox (1,1) (Board { boardSize = (5,3), boardCells = ["#####","#HH.#","#####"] })
Board { boardSize = (5,3), boardCells = ["#####","# H.#","#####"] }

cellPutBox :: Cell -> Cell

Coloca uma caixa numa célula vazia.

Exemplos de utilização:

>>> cellPutBox ' '
'H'

>>> cellPutBox '.'
'I'

cellRemoveBox :: Cell -> Cell

Retira uma caixa de uma célula com uma caixa.

Exemplos de utilização:

>>> cellRemoveBox 'H'
' '

>>> cellRemoveBox 'I'
'.'

Coordenadas

Tipos e funções relacionados com o processamento de coordenadas.

data Coords

Representa as coordenadas (posições no tabuleiro) do boneco e de todas as caixas.

Constructors

Coords

Fields

coordsPlayer :: Vector

As coordenadas do boneco.

coordsBoxes :: [Vector]

As coordenadas de todas as caixas.

parseCoords :: [String] -> Coords

Constrói um valor do tipo Coords a partir de uma representação textual adequada.

Exemplos de utilização:

>>> parseCoords ["11 2", "5 8", "7 7", "5 6"]
Coords { coordsPlayer = (11,2), coordsBoxes = [(5,8),(7,7),(5,6)] }

applyBoxesToBoard

Arguments

:: [Vector]
Coordenadas das caixas.

-> Board
Tabuleiro original.

-> Board
Tabuleiro resultante.

Modifica um tabuleiro para que nele fiquem representadas as caixas nas células em que estas se encontrem.

As células nas quais se encontram caixas passam a ter o valor:

'H' se a célula estiver vazia;

'I' se a célula for um local de arrumação.

applyBoxToBoard

Arguments

:: Vector
Coordenadas da caixa.

-> Board
Tabuleiro original.

-> Board
Tabuleiro resultante.

Modifica um tabuleiro para que nele fique representada uma caixa na célula em que esta se encontre.

A célula na qual se encontra a caixa passa a ter o valor:

'H' se a célula estiver vazia;

'I' se a célula for um local de arrumação.

Comandos

Tipos e funções relacionados com o comando que deve ser executado.

type Command = Char

Representa um comando.

Pode tomar os seguintes valores: 'L', 'R', 'U', 'D'.

parseCommands :: String -> [Command]

Contrói um valor do tipo Command a partir de uma representação textual adequada.

commandDir :: Command -> Vector

Converte um valor do tipo Command num valor do tipo Vector que representa a direção correspondente a esse comando.

Exemplos de utilização:

>>> commandDir 'L'
(-1,0)

>>> commandDir 'R'
(1,0)

>>> commandDir 'U'
(0,1)

>>> commandDir 'D'
(0,-1)

runCommands

Arguments

:: State
Estado inicial.

-> [Command]
Comandos a serem executados.

-> State
Estado após a execução dos comandos.

Executa (ou tenta executar) uma série de comandos, dado um certo estado de jogo (cf. runCommand).

runCommand

Arguments

:: State
Estado inicial.

-> Command
Comando a ser executado.

-> State
Estado após a execução do comando.

Executa (ou tenta executar) um comando, dado um certo estado de jogo. Devolve o estado após a execução do programa.

O boneco apenas se pode movimentar se:

a célula de destino estiver vazia, OU

a célula de destino contiver uma caixa mas a célula imediatamente a seguir estiver vazia.

Se não for possível mover o boneco na direção representada pelo comando, o estado não é alterado.