TarefaE

Description

Este módulo contém uma resolução correspondente à Tarefa E da segunda fase do projeto de Laboratórios de Informática I.

O objetivo é o de apresentar no stdout as dimensões do menor retângulo (alinhado com os eixos) envolvente de uma Picture, a qual é descrita pelo input fornecido através do stdin.

Entrada

tarefaE :: Picture -> (Float, Float)

Função inicial da tarefa. Recebe o input do programa e devolve o output correspondente.

Tipos e Funções Auxiliares

Tipos e funções genéricos utilizados pelo programa.

(|+|) :: Vector -> Vector -> Vector

Adiciona dois valores do tipo Vector componente por componente, segundo a equação:

(x1, y1) |+| (x2, y2) = (x1+x2, y1+y2)

Exemplos de utilização:

>>> (1,4) |+| (2,5)
(3,9)

(|*|) :: Vector -> Vector -> Vector

Multiplica dois valores do tipo Vector componente por componente, segundo a equação:

(x1, y1) |*| (x2, y2) = (x1*x2, y1*y2)

Exemplos de utilização:

>>> (1,3) |*| (2,3)
(2,9)

type Matrix = [Float]

Representa uma matrix quadrada de ordem 3.

Deve ter 9 elementos.

matGet

Arguments

:: Matrix
Matriz.

-> (Int, Int)
(Linha do elemento, Coluna do elemento)

-> Float
Elemento na posição especificada.

Devolve o elemento de uma matrix quadrada de ordem 3 na posição especificada.

Os índices da linha e da coluna devem ter valores entre 1 e 3.

Exemplos de utilização:

>>> [1,2,3,4,5,6,7,8,9] `matGet` (1,1)
1

>>> [1,2,3,4,5,6,7,8,9] `matGet` (2,1)
4

>>> [1,2,3,4,5,6,7,8,9] `matGet` (1,3)
3

>>> [1,2,3,4,5,6,7,8,9] `matGet` (3,3)
9

matMul :: Matrix -> Matrix -> Matrix

Multiplica duas matrizes quadradas de ordem 3.

Exemplos de utilização:

>>> [1,4,2,3,0,4,0,1,1] `matMul` [2,0,1,5,2,3,4,1,2]
[30,10,17,22,4,11,9,3,5]

>>> [0,0,0,0,0,0,0,0,0] `matMul` [2,0,1,5,2,3,4,1,2]
[0,0,0,0,0,0,0,0,0]

>>> [1,4,2,3,0,4,0,1,1] `matMul` [0,0,0,0,0,0,0,0,0]
[0,0,0,0,0,0,0,0,0]

>>> [1,0,0,0,1,0,0,0,1] `matMul` [2,0,1,5,2,3,4,1,2]
[2,0,1,5,2,3,4,1,2]

>>> [1,4,2,3,0,4,0,1,1] `matMul` [1,0,0,0,1,0,0,0,1]
[1,4,2,3,0,4,0,1,1]

translationMatrix

Arguments

:: Vector
Vetor da translação.

-> Matrix
Matriz resultante.

Constrói uma matriz quadrada de ordem 3 que representa uma translação no plano.

Exemplos de utilização:

>>> translationMatrix (2,3)
[1,0,2,
 0,1,3,
 0,0,1]

rotationMatrix

Arguments

:: Float
Ângulo da rotação.

-> Matrix
Matriz resultante.

Constrói uma matriz quadrada de ordem 3 que representa uma rotação no plano em relação à origem.

O ângulo deve ser especificado em radianos, no sentido contrário aos ponteiros do relógio.

Exemplos de utilização:

>>> rotationMatrix pi
[-1, 0, 0,
  0,-1, 0,
  0, 0, 1]

scalingMatrix

Arguments

:: Vector
Vetor da escala.

-> Matrix
Matriz resultante.

Constrói uma matriz quadrada de ordem 3 que representa uma transformação de escala no plano em relação à origem.

Exemplos de utilização:

>>> scalingMatrix (4,5)
[4,0,0,
 0,5,0,
 0,0,1]

data Rect

Representa um retângulo alinhado com os eixos.

Constructors

Rect

Fields

rectLeft :: Float

Coordenada, ao longo do eixo Ox, da aresta esquerda do retângulo.

rectBottom :: Float

Coordenada, ao longo do eixo Oy, da aresta inferior do retângulo.

rectRight :: Float

Coordenada, ao longo do eixo Ox, da aresta direita do retângulo.

rectTop :: Float

Coordenada, ao longo do eixo Oy, da aresta superior do retângulo.

rectSize :: Rect -> Vector

Calcula as dimensões de um retângulo representado por um valor do tipo Rect, através da equação:

rectSize (Rect l b r t) = (r-l, t-b)

Exemplos de utilização:

>>> rectSize (Rect 1 5 4 7)
(3,2)

joinRects :: [Rect] -> Maybe Rect

Calcula o menor retângulo (alinhado com os eixos) que contém todos os retângulos especificados (também alinhados com os eixos).

Se a lista de retângulos especificada for vazia, é devolvido o valor Nothing. Em caso contrário, é devolvido o valor Just rect, onde rect corresponde a um valor do tipo Rect que representa o menor retângulo que contém todos os retângulos especificados.

Exemplos de utilização:

>>> joinRects []
Nothing

>>> joinRects [Rect 1 5 4 7]
Just (Rect 1 5 4 7)

>>> joinRects [Rect 1 5 4 7, Rect (-1) 4 4 5]
Just (Rect (-1) 4 4 7)

Elementos Geométricos

Tipos e funções relacionados com a representação e manipulação de elementos geométricos no plano.

data Shape

Representa um ponto, uma linha, um polígono, uma elipse ou um conjunto desses elementos geométricos.

Funciona como tipo intermédio para o cálculo do retângulo envolvente de uma forma no plano.

Constructors

ShapePoints [Vector]
Representa um ponto, uma linha ou um polígono através de uma lista de pontos no plano. Os pontos devem representar os vértices da forma.

ShapeEllipse Matrix
Representa uma elipse através da matriz que transforma a circunferência unitária nessa elipse.

ShapeList [Shape]
Representa um conjunto de elementos geométricos.

translateShape

Arguments

:: Vector
Translação.

-> Shape
Elemento geométrico ao qual aplicar a translação.

-> Shape
Elemento geométrico resultante.

Aplica uma translação a um valor do tipo Shape.

rotateShape

Arguments

:: Float
Rotação; radianos, sentido contrário aos ponteiros do relógio.

-> Shape
Elemento geométrico ao qual aplicar a rotação.

-> Shape
Elemento geométrico resultante.

Aplica uma rotação (em torno da origem) a um valor do tipo Shape.

scaleShape

Arguments

:: Vector
Escala.

-> Shape
Elemento geométrico ao qual aplicar a transformação.

-> Shape
Elemento geométrico resultante.

Aplica uma transformação de escala a um valor do tipo Shape.

pictureToShape :: Picture -> Shape

Converte um valor do tipo Picture num valor do tipo Shape que representa o mesmo elemento geométrico no plano.

Nota: Os construtores ThickCircle, Arc, ThickArc e Text não são suportados; a função não está definida nesses casos.

shapeBounds :: Shape -> Maybe Rect

Calcula o menor retângulo (alinhado com os eixos) que contém a representação visual do valor do tipo Shape especificado.

Se o valor especificado não tiver uma representação visual associada (e.g. ShapePoints []), é devolvido o valor Nothing. Em caso contrário, é devolvido o valor Just rect, onde rect corresponde a um valor do tipo Rect que representa o menor retângulo (alinhado com os eixos) que contém a representação visual do valor do tipo Shape especificado.

pictureBounds :: Picture -> Maybe Rect

Calcula o menor retângulo (alinhado com os eixos) que contém a representação visual do valor do tipo Picture especificado.

Se o valor especificado não tiver uma representação visual associada (e.g. Blank []), é devolvido o valor Nothing. Em caso contrário, é devolvido o valor Just rect, onde rect corresponde a um valor do tipo Rect que representa o menor retângulo (alinhado com os eixos) que contém a representação visual do valor do tipo Picture especificado.

Equivalente a (shapeBounds . pictureToShape).