TarefaSqrt16

Description

Módulo om funções alternativas às da Tarefa 4, adapadas como auxiliares à componente gráfica.

Documentation

type Flames = (Int, Int, Char)

type Burning = ([String], [Flames])

avanca :: [String] -> Int -> Burning

reduce :: [String] -> [String]

Função que percorre o mapa e, a cada bomba, reduz o tepo do contador em 1.

tickBomb :: String -> String

Função que reduz o contador na string da bomba.

explosionsMap :: Burning -> [String] -> Burning

Função que percorre o mapa para encontrar bombas que têm de explodir, levando outro argumento com o mapa, que vai sendo alterado à medida que vão havendo explosões.

explode :: [String] -> String -> [String]

Função que, a partir do estado do mapa e de uma string de uma bomba prestes a explodir, gera o mapa depois da explosão.

flameUp :: [String] -> Int -> Int -> Int -> [String]

Função que explode os espaços do mapa acima da bomba e chama as funções para efetuar as alterações necessárias no etsado do jogo.

flameDown :: [String] -> Int -> Int -> Int -> [String]

Função que explode os espaços do mapa abaixo da bomba e chama as funções para efetuar as alterações necessárias no etsado do jogo.

flameRight :: [String] -> Int -> Int -> Int -> [String]

Função que explode os espaços do mapa à direita da bomba e chama as funções para efetuar as alterações necessárias no etsado do jogo.

flameLeft :: [String] -> Int -> Int -> Int -> [String]

Função que explode os espaços do mapa à esquerda da bomba e chama as funções para efetuar as alterações necessárias no etsado do jogo.

rmDoubles :: [Flames] -> [Flames]

Função que remove as coordenadas de chamas que podem entrar em conflito.

chooseFlame :: Int -> Int -> String -> Flames

makeFstrinf :: Int -> Int -> [Flames] -> String

removeFlames :: Int -> Int -> [Flames] -> [Flames]

cordFlames :: [String] -> String -> [Flames]

Funçãoq ue cria uma lista de coordenadas em qe são propagadas as chamas, que será utilizada na componente gráfica.

cordFlamesUp :: [String] -> Int -> Int -> Int -> [Flames]

cordFlamesDown :: [String] -> Int -> Int -> Int -> [Flames]

cordFlamesLeft :: [String] -> Int -> Int -> Int -> [Flames]

cordFlamesRight :: [String] -> Int -> Int -> Int -> [Flames]

espiral :: [String] -> Int -> [String]

Função que coloca as pedras e esmaga o que estiver nessas coordenadas.

cordLeft :: (Int, Int) -> Int -> [(Int, Int)]

cordRight :: (Int, Int) -> Int -> [(Int, Int)]

cordUp :: (Int, Int) -> Int -> [(Int, Int)]

cordDown :: (Int, Int) -> Int -> [(Int, Int)]

quadrado :: (Int, Int) -> Int -> [(Int, Int)]

listaSpiral :: (Int, Int) -> Int -> [(Int, Int)]

Função que ger auma lista de pares de coordenadas que respresntam a ordem pela qual o mapa será preenchido por pedras.

placeRock :: [String] -> Int -> [String]

Função que coloca a pedra relativa a um instante do jogo.

eraseObj :: [String] -> Int -> [String]

Função que "esmaga" o que estiver numas dadas coordenadas (onde será colocad auma pedra)

removePower :: [String] -> Int -> Int -> [String]

Função que remove o powerup numas dadas coordenadas.

removePlayer :: [String] -> Int -> Int -> [String]

Função que remove o jogador numas dadas coordenadas.

blowTimer :: [String] -> Int -> Int -> [String]

Função que coloca o timer da bomba numas coordenadas a 1.

stringTimer :: String -> String

Função que coloca o timer da string de uma bomba a 1 (a menos que esteja para explodir).

removeBomb :: [String] -> Int -> Int -> [String]

Função que remove o powerup numas dadas coordenadas.

isTherePower :: [String] -> Int -> Int -> Bool

Função que testa se há um powerup num dado espaço do mapa.

isTherePlayer :: [String] -> Int -> Int -> Bool

Função que testa se há um jogador num dado espaço do mapa.

isThereBomb :: [String] -> Int -> Int -> Bool

Função que testa se há um powerup num dado espaço do mapa.

pOrden1 :: String -> Int

Esta pequena função traduz a função pOrdenSt1 num Int, de forma a podermos trabalhar com esta informação em índices, por exemplo.

pAbcis1 :: String -> Int

O papel desempenhado pela pAbcis é análogo ao da pOrden, mas diz respeito à abcissa de um dado objeto (e.g. jogador, bomba, power up..)

pAbcisSt1 :: String -> String

Esta função obtem a parte da string correspondente a um jogador, ou outro objeto com coordenadas, em que se encontra a informação sobre a abcissa

das coordenadas desse mesmo objeto. Como esta informação começa sempre no 3º caracter de cada string, apenas temos de considerar o quão grande é o

número relativo a esta mesma coordenada.

pOrdenSt1 :: String -> String

Apesar de esta função ter exaamente mesmo papel que a pAbcisSt, a ordenada dos objetos vem em segundo lugar, e, por isso, os caracteres da string

que correspondem a esta informação também dependem da quantidade de caracteres utilizados para descrever a abcissa.

range :: String -> Int

Função que retorna o range da bomba descrita numa dada string.