Main

Description

Tarefa 3 da 1ª Fase do projeto de LI1.

Funções de Teste

correTestes :: IO ()

Corre múltiplos testes para a Tarefa 3.

correTeste :: ([String] -> [String]) -> String -> IO ()

Corre um teste para a Tarefa 3.

Funções da Tarefa 3

inStr :: String -> [String]

Esta função recebe uma string e transforma-a numa lista de strings, dividindo-a quando encontrar um '\n'.

outStr :: [String] -> String

Esta função recebe uma lista de strings e transforma-a numa única string, adicionando '\n' entre as strings originais.

main :: IO ()

Função geral da Tarefa 3.

tarefa3 :: [String] -> [String]

Função que recebe um tabuleiro, coordenadas e uma letra (L, U, R, D) e devolve a coordenada final do boneco.

Exemplo de utilização:

 >>> tarefa3 ["###########","#       ..#","#     #####","#     #####","#     #####","###########","1 4","4 2","4 3","D"]
 ["1 3"]

Funções Auxiliares

processa :: [String] -> [String]

Função que apaga eventuais listas vazias no final do tabuleiro.

>>> processa ["#####","#   #","##. #","#####","     "]
["#####","#   #","##. #","#####"]

parteMapa :: [String] -> ([String], [String])

Função que divide a parte das coordenadas da parte do tabuleiro.

 >>> parteMapa ["#####","# ###","#####","1 1"]
 (["#####","# ###","#####"],["1 1"])

leCoordenadas :: [String] -> [(Int, Int)]

Função que recebe a parte das coordenadas e as transforma para o formato (x,y).

 >>> leCoordenadas ["1 1","3 4"]
[(1,1),(3,4)]

Funções Principais

move :: Char -> (Int, Int) -> [(Int, Int)] -> [String] -> (Int, Int)

Ao receber L, R , U ou D, modifica a coordenada para a Esquerda, Direita, Cima ou Baixo, respetivamente, caso possível.

>>> move 'R' (1,1) [(1,1),(1,2)] ["#####","#   #","#   #","#####"]
(2,1)

Funções específicas para cada movimento

moveL :: (Int, Int) -> [(Int, Int)] -> [String] -> Bool

Esta função verifica se é possível a movimentação do boneco para esquerda. Caso seja possível, ela devolve True, e no outro caso devolve False.

>>> moveL (2,1) [(2,1),(1,2)] ["#####","#   #","#   #","#####"]
True

>>> moveL (1,1) [(1,1),(1,2)] ["#####","#   #","#   #","#####"]
False

moveR :: (Int, Int) -> [(Int, Int)] -> [String] -> Bool

Esta função verifica se é possível a movimentação do boneco para a direita. Caso seja possível, ela devolve True, e no outro caso devolve False.

>>> moveR (1,1) [(1,1),(1,2)] ["#####","#   #","#   #","#####"]
True

>>> moveR (3,1) [(3,1),(1,2)] ["#####","#   #","#   #","#####"]
False

moveU :: (Int, Int) -> [(Int, Int)] -> [String] -> Bool

Esta função verifica se é possível a movimentação do boneco para cima. Caso seja possível, ela devolve True, e no outro caso devolve False.

>>> moveU (1,1) [(1,1),(2,2)] ["#####","#   #","#   #","#####"]
True

>>> moveU (1,1) [(1,1),(1,2)] ["#####","#   #","#   #","#####"]
False

moveD :: (Int, Int) -> [(Int, Int)] -> [String] -> Bool

Esta função verifica se é possível a movimentação do boneco para baixo. Caso seja possível, ela devolve True, e no outro caso devolve False.

>>> moveD (1,2) [(1,2),(2,2)] ["#####","#   #","#   #","#####"]
True

>>> moveD (1,1) [(1,1),(1,2)] ["#####","#   #","#   #","#####"]
False

Funções que verificam se tem uma caixa ou é parede na coordenada pedida

isCaixa :: (Int, Int) -> [(Int, Int)] -> Bool

Verifica se numa determinada coordenada está uma caixa.

>>> isCaixa (2,3) [(2,3),(3,7),(4,5)]
True

>>> isCaixa (0,0) [(2,3),(3,7),(4,5)]
False

isParede :: (Int, Int) -> [String] -> Bool

Recebendo um par de coordenadas e um tabuleiro, verifica se o elemento nessa coordenada é parede ('#').

>>> isParede (0,0) ["####","#  #","####"]
True

>>> isParede (1,1) ["####","#  #","####"]
False