Main

Description

Módulo que permite a resolução de um jogo Sokoban. Este módulo recebe um input e certos comandos. O input é o mapa em si com a posição do boneco e as posições das caixas, e de seguida uma das funções gera comandos por ordem crescente de tamanho e uma outra testa se estes mesmos comandos estão corretos. No caso de um estar correto é apresentado como solução.

Documentation

type Tabuleiro = [String]

Lista com o tabuleiro.

type Coords = [String]

Lista com as coordenadas.

type Posicao = (Int, Int)

Posições (x,y) do boneco.

type Input = [String]

Input constituído pelo tabuleiro e pelas coordenadas.

type Movimento = String

String com o movimento pretendido.

type TickCount = Int

Número de movimentos efetuados.

type Comandos = String

String com os comandos, ex: ULRD.

type Linha = String

Linha do tabuleiro.

Input

main :: IO ()

Função Main.

trySolution :: Input -> Int -> [Char]

Testa as soluções geradas numa função auxiliar (comands) até encontrar a solução do jogo.

comands :: [[Char]]

Gera soluções por ordem crescente de tamanho.

jogada :: Input -> Comandos -> String

Fornce o tabuleiro à função checkMov.

checkMov :: Tabuleiro -> TickCount -> Comandos -> Posicao -> String

Fornece uma String com o tabuleiro de jogo após todas as movimentações e as implicações destas movimentações terem sido realizadas.

mvBoxAux :: Tabuleiro -> Posicao -> Movimento -> Tabuleiro

Fornece o tabuleido de jogo após o Sokoban executar um movimento juntamente com uma caixa.

existemH :: Tabuleiro -> Int

Verifica se o jogo terminou e, por isso, se todas as caixas estão nos locais de arrumação.

existemHAux :: Linha -> Int

Função auxiliar da existemH.

rmBoneco :: Tabuleiro -> Tabuleiro

Remove o Sokoban do tabuleiro.

rmBonecoAux :: Linha -> Linha

Movimentações

Verificação

move :: Tabuleiro -> Posicao -> Movimento -> Posicao

Movimentação realizada. Aqui é verificado se exite algum obstáculo do tipo de uma parede, duas caixas seguidas ou uma caixa com um muro.

Funções auxiliares

Divisão do Input

divideTab :: Input -> (Tabuleiro, Coords)

Função da Tarefa 1 (validation).

Colocação das coordenadas

Colocação geral

colocCaixas :: Tabuleiro -> [Posicao] -> Int -> Tabuleiro

Função da Tarefa 2 (visualize).

Troca por letra

colocAux :: Tabuleiro -> Posicao -> Char -> Tabuleiro

Função da Tarefa 2 (visualize).

Converte lista de duplos

stringCoorCart :: Tabuleiro -> [Posicao]

Função da Tarefa 2 (visualize).

Converte em Duplos

converterEmDuplos :: String -> Posicao

Função da Tarefa 1 (validation).