Main

Description

Módulo para uma jogada num tabuleiro de Sokoban. Este módulo recebe todas as coordenadas, o tabuleiro em si e uma lista de comandos referentes à movimentação do Sokoban, caso estas movimentações conduzam à finalização do jogo é exibido o FIM juntamente com o número de movimentos, caso os comandos não sejam suficientes para terminar o jogo é exibido INCOMPLETO com o número de movimentos.

Documentation

type Tabuleiro = [String]

Lista com o tabuleiro.

type Coords = [String]

Lista com as coordenadas.

type Posicao = (Int, Int)

Posições (x,y) do boneco.

type Input = [String]

Input constituído pelo tabuleiro e pelas coordenadas.

type Movimento = String

String com o movimento pretendido.

type TickCount = Int

Número de movimentos efetuados.

type Comandos = String

String com os comandos, ex: ULRD.

type Linha = String

Linha do tabuleiro.

type Output = [String]

Output

Input

outStr :: [String] -> String

Divide linhas do Output.

main :: IO ()

Função Main.

tarefa4 :: Input -> Output

Função geral que permite que após separar o tabuleiro o fornece à função checkMov.

checkMov :: Tabuleiro -> TickCount -> Comandos -> Posicao -> Output

Verifica se a lista dos comandos conduzes à solução do jogo.

mvBoxAux :: Tabuleiro -> Posicao -> Movimento -> Tabuleiro

Realiza o movimento de uma caixa se possível.

existemH :: Tabuleiro -> Int

Verifica se o jogo acabou. Recorre à função existemHAux.

existemHAux :: Linha -> Int

Contabiliza o número de caixas que faltam colocar.

rmBoneco :: Tabuleiro -> Tabuleiro

Remove o Sokoban do tabuleiro.

rmBonecoAux :: Linha -> Linha

Função auxiliar.

Movimentações

Verificação

move :: Tabuleiro -> Posicao -> Movimento -> Posicao

Movimentação realizada é verificação se tem algum obstáculo do tipo de uma parede, duas caixas seguidas ou uma caixa com um muro.

Funções auxiliares

Divisão do Input

divideTab :: Input -> (Tabuleiro, Coords)

Função da Tarefa 1 (validation).

Colocação das coordenadas

Colocação geral

colocCaixas :: Tabuleiro -> [Posicao] -> Int -> Tabuleiro

Função da Tarefa 2 (visualize).

Troca por letra

colocAux :: Tabuleiro -> Posicao -> Char -> Tabuleiro

Função da Tarefa 2 (visualize).

Converte lista de duplos

stringCoorCart :: Tabuleiro -> [Posicao]

Função da Tarefa 2 (visualize).

Converte em Duplos

converterEmDuplos :: String -> Posicao

Função da Tarefa 1 (validation).