Main

Description

Módulo de validação dos tabuleiros de Sokoban. Verifica o cumprimento de todas as regras e os padrões previamente definidos para os tabuleiros e as respetivas coordenadas.

De modo a optimizar o código, apenas são verificadas as coordenadas depois de verificado o tabuleiro, evitando assim tarefas desnecessárias, pois um erro no tabuleiro estará sempre antes de um erro nas coordenadas.

Além disso, existe em ambas as situações (tabuleiro e coordenadas) verificações que ocorrem linha a linha sempre que possível, ou seja, apenas é feito um conjunto de testes, uma linha de cada vez. Caso essa linha não passe nos testes não se realizam os testes para as linhas seguintes evitando assim o uso de recursos de forma desnecessária. No entanto, existem também testes que necessitam de ser feitos com o conjunto total, tanto das linhas como das coordenadas, nesta situação, caso alguma linha não passe, essa função não continua para as linhas seguintes.

Nota: É usado o codNotOk (minBound :: Int) para representar um erro sendo que, quando isto acontece, é retornada a linha onde foi detetado o erro. Quando o teste é passado é usado o codOk (maxBound :: Int).

Documentation

type Tabuleiro = [String]

Lista com o tabuleiro.

type Coords = [String]

Lista com as coordenadas.

type Posicao = (Int, Int)

Posições (x,y).

type Input = [String]

Input constituído pelo tabuleiro e pelas coordenadas.

type Linha = String

String com coordenadas ou com o tabuleiro.

Input

inStr :: String -> [String]

Recebe tabuleiro.

outStr :: [String] -> String

Divide linhas do output.

main :: IO ()

Função Main.

tarefa1 :: Input -> [String]

Função inicial.

Nesta parte da tarefa, o objetivo será confirmar se o input está correto ou se este apresenta erros. No caso de não existirem erros será apresentada a string OK. Caso existam, será apresentada a linha onde se verifica o erro.

        Input                    Output

 ###################     |         OK
 #####   ###########     |         
 #####   ###########     |         
 #####   ###########     | 
 ###      ##########     |    
 ### # ## ##########     |     
 #   # ## #####  ..#     |     
 #               ..#     |     
 ##### ### # ##  ..#     |     
 #####     #########     |         
 ###################     |         
 11 2
 5 8
 7 7
 5 6
 7 6
 2 3
 5 3

        Input                    Output

 ###################     |         7 
 ##                #     |         
 #####      ##   ###     |         
 ###### #   ##   ###     |       
 ###################     |
 5 3
 7 3
 7 3

Divisão do Input em tabuleiro e coordenadas

divideTab :: Input -> (Tabuleiro, Coords)

Divide o input em tabuleiro e coordenadas.

Exemplo de utilização

>>> divideTab ["####","#  #","12 2","1 15"]
(["####","#  #"],["12 2","1 15"])

Verificação geral

validatab :: [String] -> String

Validação geral (tabuleiro e coordenadas).

codOk :: Int

Código a retornar caso OK.

codNotOk :: Int

Código a retornar em caso de erro.

Validação linha a linha do tabuleiro

validalinhastab :: Input -> Int

Validação das linhas.

validaLinha :: Tabuleiro -> Int -> Int -> Int

Validação individual de cada linha.

Validação linha a linha das coordenadas

validacoordtab :: Input -> Int

Validação geral das coordenadas.

validaCoordenadas :: Coords -> Tabuleiro -> Int -> Int -> Int -> Int

Validação linha a linha das coordenadas.

Funções de verificação do tabuleiro

Verifica o comprimento do tabuleiro

verificaComprimento :: Linha -> Int -> Int

Verifica se o comprimento das linhas do tabuleiro é igual ao comprimento da head do tabuleiro.

Exemplo de utilização

>>> verificaComprimento "####" 4
9223372036854775807 
>>> verificaComprimento "#" 3
-9223372036854775808

Verifica paredes (topo e base)

verificaParedesTopBot :: Linha -> Int

Verifica a existência de "paredes" na parte superior e inferior do tabuleiro, ou seja, confirma se todos os elementos da primeira e última linha do tabuleiro são um "#".

Exemplos de utilização

>>> verificaParedesTopBot "####"
9223372036854775807
>>> verificaParedesTopBot "## .#"
-9223372036854775808

Verifica paredes laterais

verificaParedesSide :: Linha -> Int

Verifica a existência de "paredes" laterais, ou seja, confirma se o primeiro e último elemento de cada linha do tabuleiro é um "#".

Exemplos de utilização

>>> verificaParedesSide "#  #"
9223372036854775807
>>> verificaParedesSide "# #."
-9223372036854775808

Verifica conteúdo

verificaContent :: Linha -> Int

Testa se o conteúdo das linhas do tabuleiro é diferente do esperado, isto significa que as linhas apenas podem ser constituídas pelos caractéres: ''#'', '' '' e ''.''

Exemplos de utilização

>>> verificaContent "# #"
 9223372036854775807
>>> verificaContent "#as#"
 -9223372036854775808

Funções de verificação das coordenadas

Contagem de pontos

contagemPontos :: Tabuleiro -> Int

Contagem do número total de pontos no tabuleiro (locais de arrumação) que são posteriormente comparados com o número total de coordenadas.

Exemplos de utilização

>>> contagemPontos ["##.","#..#"]
3

contagemPontosAux :: Linha -> Int

Função auxiliar da contagemPontos.

Compara coordenadas com pontos

confirmNumberOfCoor :: Coords -> Tabuleiro -> Int -> Int

Confirma o número de coordenadas inseridas e compara com o número de locais de arrumação.

Exemplos de utilização

>>> confirmNumberOfCoor ["12 2","13 3"] ["##.","##"] 4
9223372036854775807
>>> confirmNumberOfCoor ["12 2","13 3"] ["#..","##"] 4
5
>>> confirmNumberOfCoor ["12 2","13 3"] ["#  ","##"] 4
4

Separa string com coordenadas em duas strings

listaParaDuplo :: [String] -> (String, String)

Recebe uma lista e retorna um par. Nota: de forma a evitar erros de tipo, quando a length da lista é diferente de 2, retorna o duplo (-1,-1) que seguidamente retorna erro quando é verificada na função confValuesCoor.

Exemplos de utilização

>>> listaParaDuplo ["12","2"]
("12","2")
>>> listaParaDuplo ["12","1","3"]
("-1","-1")

Verifica se recebe números

verificaValorCoor :: Linha -> Int

Verifica se as coordenadas são números.

Exemplos de utilização

>>> verificaValorCoor "12 2"
9223372036854775807
>>> verificaValorCoor "12 a"
-9223372036854775808

Converter coordenadas em duplos

converterEmDuplos :: Linha -> Posicao

Converte coordenadas em duplos de inteiros. Nota: As coordenadas apenas serão verificadas no caso destas serem números para evitar erros de tipo.

Exemplos de utilização

>>> converterEmDuplos "3 4"
(3,4)

Verifica intervalo de localização no tabuleiro

confValuesCoor :: Posicao -> Int -> Int -> Int

Verifica se as coordenadas estão no interior do tabuleiro e se são positivas.

Exemplos de utilização

>>> confValuesCoor (3,3) 10 12
9223372036854775808
>>> confValuesCoor (3,3) 2 3
-9223372036854775808
>>> confValuesCoor (-3,0) 1 2
-9223372036854775808

confRepCoor :: Coords -> [String] -> Int -> Int

Confirma se existem coordenadas repetidas.

Exemplos de utilização

>>> confRepCoor ["12 2","2 1","12 2"] [] 17
20
>>> confRepCoor ["1 2","3 2","1 5"] [] 7
9223372036854775807

Confirma local

confColocC :: Tabuleiro -> Posicao -> Int

Confirma se é possível colocar as coordenadas no tabuleiro e que estas não se sobrepõem num muro.

Exemplos de utilização

>>> confColocC ["####","####","## #","####"] (2,1)
9223372036854775807
>>> confColocC ["####","## #","####","####"] (1,1)
-9223372036854775808