Main

Description

Módulo para uma jogada num tabuleiro de Sokoban. Este módulo recebe todas as coordenadas e o tabuleiro em si e retorna a posição do boneco após a movintação, caso esta não seja posível retorna a posição atual.

Documentation

type Tabuleiro = [String]

Lista com o tabuleiro.

type Coords = [String]

Lista com as coordenadas.

type Posicao = (Int, Int)

Posições (x,y).

type Output = String

Output com a posição final.

type Input = [String]

Input constituído pelo tabuleiro e pelas coordenadas.

type Movimento = String

String com o movimento pretendido.

Input

outStr :: [String] -> String

Divide linhas do output.

main :: IO ()

Função Main.

tarefa3 :: Input -> [Output]

Função geral.

Nesta parte da tarefa, o objetivo será determinar qual a posição final do boneco após a execução de um movimento. Encontram-se abaixo indicados exemplos do resultado final.

        Input                     Output

 #####################     |       8 2
 #### ..  #          #     |
 ####               ##     |
 ###     #           #     |
 #####################     |
 8 2
 8 3
 3 1
 D
        Input                     Output

 #####################     |       12 2
 #### ..             #     |        
 ####               ##     |        
 ###                 #     |        
 #####################     |        
 12 1
 8 3
 3 1
 U

Movimentações

Realiza movimentação

play :: Tabuleiro -> Output

Movimentação realizada num tabuleiro onde estão colocadas as caixas e as paredes.

Verificação

move :: Tabuleiro -> Posicao -> Movimento -> Posicao

É verificado se existe algum obstáculo, isto é, uma parede, duas caixas seguidas ou uma caixa com um muro.

Exemplo de utilização

>>> move ["#####","#.  #","#####"] (3,1) "L"
"2 1"

Print Output

output :: Posicao -> Output

Converte par ordenado em string que separa os número com os espaços.

Exemplo de utilização

>>> output (3,2)
"3 2"

Funções auxiliares

Divisão do Input

divideTab :: Input -> (Tabuleiro, Coords)

Função da Tarefa 1 (validation).

Colocação das coordenadas

Colocação geral

colocElem :: Tabuleiro -> [Posicao] -> Int -> Tabuleiro

Função da Tarefa 2 (visualize).

Troca por letra

colocAux :: Tabuleiro -> Posicao -> Char -> Tabuleiro

Função da Tarefa 2 (visualize).

Converte lista de duplos

stringCoorCart :: Tabuleiro -> [Posicao]

Função da Tarefa 2 (visualize).

Converte em Duplos

converterEmDuplos :: String -> Posicao

Função da Tarefa 1 (validation).