Main

Description

Módulo que permite determinar as dimensões do menor retângulo envolvente de uma Picture. Este módulo recebe uma Picture e retorna a altura e a largura respetivamente do retângulo formado em torno da Picture.

Documentation

type Degree = Float

Ângulo em graus.

type Radian = Float

Ângulo em radianos.

type Coord = (Float, Float)

Coordenada do tipo (x,y).

type CoordPolar = (Float, Radian)

Coordenada polar do tipo (d , alpha).

Input

main :: IO ()

Função Main.

tarefa5 :: Picture -> (Int, Int)

Função que retorna a altura e a largura respetivamente do retângulo formado em torno de uma Picture.

Calcular as coordenadas dos pontos do rectângulo que envolve cada Picture

coords :: Picture -> [Coord]

Função que calcula individualmente os pontos do rectângulo que envolve cada uma das Picture.

separa :: Picture -> [Coord]

Caso o input receba um Pictures, a função separa devolve uma lista com as diferentes coordenadas do de cada retângulo envolvente da _Picture__.

calcExtrem :: (Ord a, Ord b) => (Num a, Num b) => [(a, b)] -> [(a, b)]

Função que fornece as coordenadas dos extremos (extremo superior esquerdo e inferior direito respetivamente) de um retângulo formado em volta de uma Picture.

scalePic :: Float -> Float -> [Coord] -> [Coord]

Função que determina as coordenadas finais após um Scale.

translatePic :: Float -> Float -> Picture -> [Coord]

Função que determina as coordenadas finais após um Translate x y.

rotation :: Degree -> [Coord] -> [Coord]

Função que efetua a rotação de um determinado ponto (x,y) segundo um dado grau. A rotação é efetuada no sentido horário.

joinLists :: [a] -> [b] -> [(a, b)]

Junção de duas listas, uma lista com o x e outra com os y, resultando uma lista com pares ordenados (x,y).

pointsCircle :: Float -> [(Float, Float)]

Calcula pontos de um círculo com um determinado intervalo entre eles.

Conversão de ângulos e conversão de coordenadas

degToRad :: Degree -> Radian

Converte graus em radianos.

radToDeg :: Radian -> Degree

Converte radianos em graus.

auxToPolar :: Coord -> CoordPolar

Converte um ponto para coordenada polar, do tipo ( d, alpha ), onde d é distância do ponto à origem e alpha o ângulo formado.

auxToRet :: CoordPolar -> Coord

Converte um ponto para coodenadas retangulares.

Testes HUnit

tarefa5test :: Test

Testar a área de uma Picture com recurso ao HUnit.