Main

Description

Tarefa 4 da 2ª Fase do projeto de LI1.

Funções da Tarefa 4

inStr :: String -> [String]

Esta função recebe uma string e transforma-a numa lista de strings, dividindo-a quando encontrar um '\n'.

outStr :: [String] -> String

Esta função recebe uma lista de strings e transforma-a numa única string, adicionando '\n' entre as strings originais.

main :: IO ()

Função geral da tarefa 4.

tarefa4 :: [String] -> [String]

Função que recebe um tabuleiro, coordenadas e uma sequência de comandos (L, U, R, D) e devolve "INCOMPLETO <tick_count>" se o puzzle não ficar resolvido ou "FIM <tick_count>" se o puzzle for completado. <tick_count> é o número de movimentos efetuados.

Exemplo de utilização:

 >>> tarefa4 ["###########","#       ..#","#     #####","#     #####","#     #####","###########","1 4","4 2","4 3","DDRRRDRUUDLLUURRRRRLLLDDLULURRRR"]
 ["FIM 32"]

Funções Principais

movegeral :: [String] -> (Int, [(Int, Int)])

A função movegeral recebe uma lista de strings e processa-a de modo a que esta possa ser utilizada pela função move.

move :: String -> (Int, Int) -> [(Int, Int)] -> [String] -> [(Int, Int)] -> (Int, [(Int, Int)])

A função move recebe os comandos, as coordenadas do boneco, as coordenadas das caixas, o tabuleiro e as coordenadas dos lugares de arrumação e devolve um tuplo, cujo primeiro elemento diz respeito ao número de movimentos efetuados e o segundo é uma lista com as coordenadas finais das caixas.

moveM :: (Int, Int) -> (Int, Int) -> [(Int, Int)] -> [String] -> Bool

Verifica se um determinado movimento é possível.

moveCaixa :: (Int, Int) -> (Int, Int) -> [(Int, Int)] -> [(Int, Int)]

Conforme a posição anterior e a atual do boneco verifica se houve movimento de uma caixa e devolve uma lista com as coordenadas novas das caixas.

coordsarrumacao :: [String] -> [(Int, Int)]

Recebe o tabuleiro e devolve as coordenadas dos lugares de arrumação.

processa :: [String] -> [String]

Função que apaga eventuais listas vazias no final do tabuleiro.

parteMapa :: [String] -> ([String], [String])

Função que divide a parte das coordenadas da parte do tabuleiro.

leCoordenadas :: [String] -> [(Int, Int)]

Função que recebe a parte das coordenadas e as transforma para o formato (x,y).

 >>> leCoordenadas ["1 1","3 4"]
[(1,1),(3,4)]

equiv :: [(Int, Int)] -> [(Int, Int)] -> Bool

Verifica se duas listas de coordenadas são iguais.

isCaixa :: (Int, Int) -> [(Int, Int)] -> Bool

Verifica se numa determinada coordenada está uma caixa.

>>> isCaixa (2,3) [(2,3),(3,7),(4,5)]
True

>>> isCaixa (0,0) [(2,3),(3,7),(4,5)]
False

isParede :: (Int, Int) -> [String] -> Bool

Recebendo um par de coordenadas e um tabuleiro, verifica se o elemento nessa coordenada é parede ('#').

>>> isParede (0,0) ["####","#  #","####"]
True

>>> isParede (1,1) ["####","#  #","####"]
False