Main

Description

Tarefa 6 da 2ª Fase do projeto de LI1.

Documentation

main :: IO ()

Função principal que invoca o jogo.

Função que cria o jogo.

joga :: mundo -> (mundo -> IO Picture) -> (Event -> mundo -> IO mundo) -> (Float -> mundo -> mundo) -> IO ()

Esta função interage com todos os componentes do jogo.

data Mapa

Define o tipo de Mapa, permitindo o uso de Records.

Constructors

Mapa

Fields

score :: (Int, Int)

memoria :: [[(Float, Float)]]

nome :: String

imagens :: [Picture]

coord :: [(Float, Float)]

mapi :: [String]

nivel :: Int

tempo :: (Int, Int)

sh :: Int

Função que desenha o mapa

desenhaMapa :: Mapa -> IO Picture

Transforma o Mapa numa IO Picture, ou seja, na sua representação gráfica.

Função que reage a um evento (carregar numa tecla, mover o rato, etc)

reageEvento :: Event -> Mapa -> IO Mapa

Reage ao pressionar das setas do teclado, modificando o Mapa, e por consequência, a representação gráfica.

Funções que controlam o movimento do boneco e das caixas.

moveBoneco :: Char -> Mapa -> Mapa

Esta função recebe um movimento e um Mapa e retorna o Mapa resultante desse movimento.

move :: Char -> [(Float, Float)] -> [String] -> [(Float, Float)]

A função move recebe um comando, as coordenadas do boneco e das caixas e o tabuleiro e devolve as coordendas do boneco e da caixas depois das alterações resultantes desse comando.

moveCaixa :: (Float, Float) -> (Float, Float) -> [(Float, Float)] -> [(Float, Float)]

Conforme a posição anterior e a atual do boneco verifica se houve movimento de uma caixa e devolve uma lista com as coordenadas novas das caixas.

moveM :: (Float, Float) -> (Float, Float) -> [(Float, Float)] -> [String] -> Bool

Verifica se um determinado movimento é possível.

Função que permite fazer restart.

restart :: Mapa -> Mapa

Esta função faz o jogo voltar ao seu estado inicial.

Função que permite fazer undo.

undo :: Mapa -> Mapa

Esta função faz o jogo retroceder uma jogada.

Função que permite escrever o nome.

name :: Char -> Mapa -> Mapa

Esta função permite escrever o nome se obtiver-se uma melhor pontuação.

Funções que permitem avançar de nivel.

nextlevel :: Mapa -> IO Mapa

Constrói o nível seguinte, exceto de se estiver no último.

previouslevel :: Mapa -> IO Mapa

Constrói o nível anterior, exceto de se estiver no primeiro.

constroimapa :: Mapa -> String -> IO Mapa

Atravês de um Mapa e um ficheiro, contrói um novo nível de acordo com a informação desse ficheiro.

Função que permite fazer shuffle.

shuffle :: Mapa -> IO Mapa

A função escolhe aleatoriamente novos locais para as caixas, preservando aquelas que já estão num local de arrumação.

Função que reage ao passar do tempo

reageTempo :: Float -> Mapa -> Mapa

Esta função permite contabilizar oo tempo.

Funções auxiliares

inStr :: String -> [String]

Esta função recebe uma string e transforma-a numa lista de strings, dividindo-a quando encontrar um '\n'.

coordenadas :: Char -> [String] -> [(Float, Float)]

Recebe um caractere e um tabuleiro e devolve as coordenadas onde ocorre aquele caractere.

processaMapa :: String -> ([String], [(Float, Float)])

Função que recebe um ficheiro e processa-o de modo a ficar com um tuplo cujo o primeiro elemento é o tabuleiro e o segundo as coordenadas.

parteMapa :: [String] -> ([String], [String])

Função que divide a parte das coordenadas da parte do tabuleiro.

leCoordenadas :: [String] -> [(Float, Float)]

Função que recebe a parte das coordenadas e as transforma para o formato (x,y).

 >>> leCoordenadas ["1 1","3 4"]
[(1,1),(3,4)]

rmov :: [String] -> [String]

Retira os caracteres '#' que forem redundantes.

>>>  rmov  ["##########","##     ###","## ###   #","#        #","# ..#   ##","##..#   ##","##########"] ["##########","##     ###","## ###   #","#        #","# ..#   ##","##..#   ##","##########"]
[" #######  "," #     ###","## ###   #","#        #","# ..#   ##","##..#   # "," ######## "]

remove2 :: (Eq a, Num a) => Char -> (a, Int) -> [[Char]] -> [Char]

Verifica se a remoção do caractere '#' é possível. Caso seja possível, remove-o; caso contrário, mantém-no.

>>>  remove2 '#' (1,1) ["#####","## ##","#####"] 
"#"

>>>  remove2 '#' (0,0) ["#####","## ##","#####"] 
" "

equiv :: [(Float, Float)] -> [(Float, Float)] -> Bool

Verifica se duas listas de coordenadas são iguais.

maislindo :: String -> String

Adiciona '0' à esquerda de um número para uniformizar a sua apresentação.

concluido :: Mapa -> Bool

Verfica se o jogo está completo, ou seja , concluido.

worst :: Mapa -> Bool

Verifica se teve uma pior prestação que um jogador anterior.

better :: Mapa -> Bool

Verifica se teve uma melhor prestação que qualquer jogador anterior.

tarefa1 :: IO Picture

Cria a parte da interface gráfica que não se altera no decorrer do jogo.

tarefa2 :: [String] -> Int -> IO Picture

Cria a parte da interface gráfica que apenas só se altera quando se muda de nivel.