Main

Description

Módulo adapatado da tarefa 4 e que funciona como esqueleto do jogo Sokoban. Este módulo diz respeito à movimentação do Sokoban. Após receber o Input e os comandos necessários à movimentação do Sokoban, a função move fornece um tabuleiro com a posição final do Sokoban após essa movimentação.

Documentation

type Tabuleiro = [String]

Lista com o tabuleiro.

type Coords = [String]

Lista com as coordenadas.

type Posicao = (Int, Int)

Posições (x,y) do boneco.

type Input = [String]

Input constituído pelo tabuleiro e pelas coordenadas.

type Movimento = String

String com o movimento pretendido.

type TickCount = Int

Número de movimentos efetuados.

type Comandos = String

String com os comandos, ex: ULRD.

type Linha = String

Linha do tabuleiro.

Input

outStr :: [String] -> String

main :: IO ()

moves :: Input -> Comandos -> Tabuleiro

Função principal do movimento do Sokoban.

checkMov :: Tabuleiro -> TickCount -> Comandos -> Posicao -> Tabuleiro

Exibe o tabuleiro de jogo após todas as movimentações e as implicações destas movimentações terem sido realizadas.

mvBoxAux :: Tabuleiro -> Posicao -> Movimento -> Tabuleiro

Fornece o tabuleido de jogo após o Sokoban executar um movimento juntamente com uma caixa.

bomba :: Tabuleiro -> Posicao -> Tabuleiro

No caso de o jogador poder usar a bomba, a função remove paredes num raio de 1 caixa.

removeWall :: Tabuleiro -> Posicao -> Tabuleiro

Função auxiliar da função bomba que remove as parede envolventes do Sokoban num raio de uma caixa.

addOutsideWall :: Tabuleiro -> Tabuleiro

Adiciona uma parede à volta do mapa para ser possível utilizar a bomba. Neste caso permite a permanência de uma parede.

existemH :: Tabuleiro -> Int

Verifica se o jogo terminou e, por isso, se todas as caixas estão nos locais de arrumação.

existemHAux :: Linha -> Int

Função auxiliar da existemH.

rmBoneco :: Tabuleiro -> Tabuleiro

Remove o Sokoban do tabuleiro.

rmBonecoAux :: Linha -> Linha

Função auxiliar.

Movimentações

Verificação

move :: Tabuleiro -> Posicao -> Movimento -> Posicao

Movimentação realizada é verificação se tem algum obstáculo do tipo de uma parede, duas caixas seguidas ou uma caixa com um muro.

Funções auxiliares

Divisão do Input

divideTab :: Input -> (Tabuleiro, Coords)

Função da Tarefa 1 (validation).

Colocação das coordenadas

Colocação geral

colocCaixas :: Tabuleiro -> [Posicao] -> Int -> Tabuleiro

Função da Tarefa 2 (visualize).

Troca por letra

colocAux :: Tabuleiro -> Posicao -> Char -> Tabuleiro

Função da Tarefa 2 (visualize).

Converte lista de duplos

stringCoorCart :: Tabuleiro -> [Posicao]

Função da Tarefa 2 (visualize).

Converte em Duplos

converterEmDuplos :: String -> Posicao

Função da Tarefa 1 (validation).

Remoção de '#' "redundantes"

Marcação de redundantes

transformCar :: Tabuleiro -> Posicao -> Tabuleiro

Depois de verificados se redundantes, os cardinais que o sejam, são transformados em '@' para serem posteriormente transformados em espaços. Os que não forem redundantes continuam inalterados.

Exemplos de utilização

>>> transformCar ["########","####   #","####   #","########"] (0,0)
["@@@#####","@@@#   #","@@@#   #","@@@#####"]

Verifica redundância

checkCar :: Tabuleiro -> Posicao -> Bool

Verifica se o cardinal é redundante ou não, ou seja, verifica se a toda à volta existem apenas cardinais ou arrobas (@ que correspondem a cardinais já marcados como redundantes).

Exemplos de utilização

>>> checkCar ["@@@#####","@@##   #","####   #","########"] ( 1 , 2 )
True