Main

Description

Tarefa 5 da 2ª Fase do projeto de LI1.

Documentation

main :: IO ()

Função geral da tarefa E.

tarefa5 :: Picture -> String

Função que dada uma picture, devolve um par de floats (largura, altura) com as dimensões do menor retângulo envolvente dessa picture.

Exemplo de utilização:

 >>> tarefa5 (Circle 20)
"40 40"

Funções Secundárias

poli :: Path -> (Float, Float)

Função que dada uma lista de pontos devolve um par de floats. O primeiro corresponde à diferença entre a abcissa maior e abcissa menor e o segundo corresponde à diferença entre a ordenada maior e a ordenada menor.

tarefa :: Picture -> Path

Função que recebe uma picture e a transforma numa lista de pontos.

circulo :: Float -> Point -> Path

A partir do raio e de um centro cria uma lista de pontos que formam um círculo.

rad :: Float -> Float

Converte graus para radianos.

 >>> rad (-90)
1.5707964

 >>> rad 60
-1.0471976