Main

Description

Módulo para realização da interface gráfica do jogo. Para isso utilizámos a ferramenta Gloss. Esta tarefa funciona quase como um tradutor para Bitmaps e permite a visualização do jogo.

Documentation

type Tabuleiro = [String]

Lista com o tabuleiro.

type Coords = [String]

Lista com as coordenadas.

type Posicao = (Int, Int)

Posições (x,y) do boneco.

type Bitmaps = [Picture]

Lista com todos os bitmaps.

type NivelMapa = String

Tabuleiro numa string apenas.

type Comandos = String

String com os comandos, ex: ULRD.

type ComandoSimples = Char

Apenas um comando, ex: U.

type Bomba = Bool

Existência de uma bomba ou não.

type Contador = Float

Contador do tipo Float.

type ContadorInt = Int

Contador do tipo Int.

type Level = String

Número de um nível, ex: ”3”.

type Tema = String

Nome de um tema, ex: ”Lego”.

type Highscore = String

Melhor pontuação obtida no jogo.

type Mapa = (Bitmaps, NivelMapa, Comandos, Bomba)

Base do jogo.

Input

main :: IO ()

Função Main.

joga :: world -> (world -> Picture) -> (Event -> world -> world) -> IO ()

Função fornecida.

Interface do Jogo Sokoban

desenhaMapa :: Mapa -> Picture

Função que permite desenhar a interface do jogo. Consiste assim na parte visual.

Mapa

mapshow :: Tabuleiro -> Bitmaps -> Comandos -> Picture

Função que permite após receber um tabuleiro de jogo, convertê-lo num conjunto de pictures.

translateLines :: [Picture] -> Contador -> [Picture]

Após todo o tabuleiro ser transcrito para Bitmaps, posiciona-se linha a linha o mapa.

mapshowaux :: Tabuleiro -> Bitmaps -> Comandos -> [Picture]

Converte o mapa para bitmaps correspondentes a cada símbolo.

bombshow :: Tabuleiro -> Bitmaps -> Comandos -> [Picture]

Em caso da bomba ser ativa mostra-a no tabuleiro.

assignbmp :: Contador -> Bitmaps -> NivelMapa -> Comandos -> [Picture]

Traduz linhas do tabuleiro para uma lista de picture com os bitmaps correspondentes. Ficam automaticamente posicionados relativamente ao eixo Ox.

placeBomb :: Contador -> Bitmaps -> NivelMapa -> Comandos -> [Picture]

Coloca a bomba em cima do boneco.

Ligação entre a interface gráfica e a movimentação

regEv :: Event -> Mapa -> Mapa

Converte o Output do reageEvento de IO Mapa para Mapa.

reageEvento :: Event -> Mapa -> IO Mapa

Fornece à função makeMov o comando correspondente à movimentação.

makeMov :: Mapa -> ComandoSimples -> IO Mapa

Permite ser vísivel a movimentação do Sokoban.

restart :: Mapa -> Contador -> IO Mapa

Permite ao jogador a funcionalidade de recomeçar a jogar.

undo :: Mapa -> IO Mapa

Permite ao jogador a funcionalidade de voltar à jogada anterior.

Mudar de níveis

nextLvl :: Mapa -> IO Mapa

Função que permite ao jogador a hipótese de seguir para o próximo nível.

prevLvl :: Mapa -> IO Mapa

Função que permite ao jogador a hipótese de recuar um nível.

Mudar de temas

nextTheme :: Mapa -> IO Mapa

Função que permite ao jogador a hipótese de mudar de tema. No caso de alterar de tema, todos os Bitmaps são substituídos.

reageTempo :: Float -> world -> world

Função que reage com o tempo.

Ficheiro Settings

settings :: t -> [String]

Para uma melhor organização, o ficheiro settings.txt contém toda a informação relativa aos níveis e temas. Sempre que estes dois são alterados a função settings altera o ficheiro settings.txt.

settingsFile :: t -> IO [String]

maxLevel :: Level

Função quer permite identificar o número máximo de níveis presentes na pasta /content/levels.

lvl :: t -> Level

Função que fornece o current level.

level :: t -> Level

Fornece o path para o nível.

writeLvl :: Char -> IO ()

Permite modificar no ficheiro settings.txt o nível que o jogador está a jogar.

changelvl :: Char -> [Level]

Se o jogador alterar de nível, esta função altera o level no ficheiro settings.txt.

resetLvl :: t -> IO ()

Função que altera o level no ficheiro settings.txt para 1.

theme :: t -> Tema

Função que lê qual o tema a ser usado.

themeName :: t -> Tema

Função que fornece o tema a ser usado no jogo.

newTheme :: t -> IO ()

Em caso de o jogador alterar o tema, este modificado no ficheiro settings.txt.

changeTheme :: t -> ContadorInt -> [Tema]

Permite alterar o tema do jogo.

updateSettings :: IO ()

Função que permite atualizar o ficheiro settings.txt.

updateThemes :: IO ()

Escreve todos os temas disponiveis no ficheiro settings.txt.

allThemes :: [Tema]

Fornece uma lista com todos os temas presentes do diretório content/bitmaps/themes.

updateLevels :: IO ()

Escreve todos os níveis disponiveis no ficheiro settings.txt.

listLevels :: [Level]

Fornece uma lista com todos os níveis disponíveis.

HighScore

currentHighscore :: t -> Highscore

Função que permite saber o highscore de uma determinado nível.

highscore :: Int -> IO ()

Compara e altera o ficheiro highscore do nível correspondente no caso de ser obtido um melhor score.

scores :: t -> IO Highscore

Permite a leitura do highscore do nível correspondente.

endgame :: NivelMapa -> Bool

Testa se o jogo terminou. Utiliza a função existemH para testar.

existemH :: String -> Int

Verifica se o jogo terminou e, por isso, se todas as caixas estão nos locais de arrumação.

rmBombListMoves :: String -> String

Remove o movimento B dos comandos, ou seja, usar a Bomba não conta como movimento.