Main

Description

Módulo de "vizualização" dos tabuleiros de Sokoban. Este módulo faz a remoção dos caracteres '#' redundantes e faz a coloção do boneco e das caixas no tabuleiro. Retorna o tabuleiro após estas alterações realizadas.

Documentation

type Tabuleiro = [String]

Lista com o tabuleiro.

type Coords = [String]

Lista com as coordenadas.

type Posicao = (Int, Int)

Posições (x,y).

type Input = [String]

Input constituído pelo tabuleiro e pelas coordenadas.

Input

outStr :: [String] -> String

Divide linhas do Output.

main :: IO ()

Função Main.

tarefa2 :: Input -> Tabuleiro

Função geral.

Nesta parte da tarefa, o objetivo será produzir um mapa mais simplificado, onde serão removidos os cardinais redundantes e ainda adicionadas as caixas e o boneco. Encontram-se abaixo indicados dois exemplos do resultado final.

        Input                 Tabuleiro simplificado

 ###################     |         #####
 #####   ###########     |         #   #
 #####   ###########     |         #H  #
 #####   ###########     |       ###  H##
 ###      ##########     |       #  H H #
 ### # ## ##########     |     ### # ## #   ######
 #   # ## #####  ..#     |     #   # ## #####  ..#
 #               ..#     |     # H  H          ..#
 ##### ### # ##  ..#     |     ##### ### #o##  ..#
 #####     #########     |         #     #########
 ###################     |         #######
 11 2
 5 8
 7 7
 5 6
 7 6
 2 3
 5 3

        Input                 Tabuleiro simplificado

 ################        |            ########
 #########      #        |            #      #
 #########  #  ##        |         ####  #  ##
 ######        ##        |         #      H #
 ##########    ##        |         #####    #
 #########   # ##        |    #########  H# ##
 #.     ##      #        |    #.     ## H    #
 ##.            #        |    ##.   o        #
 #.     #########        |    #.     #########
 ################        |    ########
 6 2
 10 3
 11 4
 12 6

Remoção de '#' "redundantes"

Verifica redundância

checkCar :: Tabuleiro -> Posicao -> Bool

Verifica se o cardinal é redundante ou não, ou seja, verifica se a toda à volta existem apenas cardinais ou arrobas (@ que correspondem a cardinais já marcados como redundantes).

Exemplos de utilização

>>> checkCar ["@@@#####","@@##   #","####   #","########"] ( 1 , 2 )
True

Marcação de redundantes

transformCar :: Tabuleiro -> Posicao -> Tabuleiro

Depois de verificados se são redundantes, os cardinais são transformados em '@' para serem posteriormente transformados em espaços. Os que não forem redundantes continuam inalterados.

Exemplos de utilização

>>> transformCar ["########","####   #","####   #","########"] (0,0)
["@@@#####","@@@#   #","@@@#   #","@@@#####"]

Redundantes para espaço

rmCar :: Tabuleiro -> Tabuleiro

Após a validação, os '#' redundantes (transformados em '@') são substituídos por espaços.

Exemplos de utilização

>>> rmCar ["@@@#####","@@@#   #","@@@#   #","@@@#####"]
["   #####","   #   #","   #   #","   #####"]

Colocação das coordenadas

Colocação geral

colocElem :: Tabuleiro -> [Posicao] -> Int -> Tabuleiro

Função que marca no tabuleiro as posições do boneco e das caixas. Caso seja a primeira vez a correr a função, a posição do boneco corresponde ao 'o' , senão é verificada se a posição é um ponto. Neste caso é colocado um 'I' correspondente à posição final de uma caixa, ou um 'H' (caixa em posição não-final).

Exemplos de utilização

>>> colocElem ["########","####  .#","####  .#","########"] [(4,1),(4,2),(6,2)] 1
["########","####H I#","####o .#","########"]

Troca por Letra

colocAux :: Tabuleiro -> Posicao -> Char -> Tabuleiro

Efetua-se a troca de um lugar (coordenada) pela letra correspondente mencionada anteriormente.

Exemplos de utilização

>>> colocAux ["########","####  .#","####o .#","########"] ( 4 , 2 ) 'H'
["########","####H .#","####o .#","########"]

Funções auxiliares

Divisão do Input

divideTab :: Input -> (Tabuleiro, Coords)

Função da Tarefa 1 (validation).

Converte lista de duplos

stringCoorCart :: Tabuleiro -> [Posicao]

Converte as coordenadas numa lista de duplos de inteiros.

Exemplos de utilização

>>> stringCoorCart ["4 1","4 2","6 2"]
[(4,1),(4,2),(6,2)]

Converte em duplos

converterEmDuplos :: String -> Posicao

Função da Tarefa 1 (validation).