Main

Description

Tarefa 2 da 1ª Fase do projeto de LI1.

Funções de Teste

correTestes :: IO ()

Corre múltiplos testes para a Tarefa 2.

correTeste :: ([String] -> [String]) -> String -> IO ()

Corre um teste para a Tarefa 2.

Funções da Tarefa 2

inStr :: String -> [String]

Esta função recebe uma string e transforma-a numa lista de strings, dividindo-a quando encontrar um '\n'.

outStr :: [String] -> String

Esta função recebe uma lista de strings e transforma-a numa única string, adicionando '\n' entre as strings originais.

main :: IO ()

Função geral da Tarefa 2.

tarefa2 :: [String] -> [String]

Esta função recebe um tabuleiro e coordenadas e devolve o tabuleiro no formato pedido.

Exemplo de utilização:

 >>> tarefa2 ["###########","#       ..#","#     #####","#     #####","#     #####","###########","1 4","4 2","4 3"]
 ["###########","#o      ..#","#   H #####","#   H #    ","#     #    ","#######    "]

Funções Auxiliares

processa :: [String] -> [String]

Função que apaga eventuais listas vazias no final do tabuleiro.

>>> processa ["#####","#   #","##. #","#####","     "]
["#####","#   #","##. #","#####"]

parteMapa :: [String] -> ([String], [String])

Função que divide a parte das coordenadas da parte do tabuleiro.

 >>> parteMapa ["#####","# ###","#####","1 1"]
 (["#####","# ###","#####"],["1 1"])

leCoordenadas :: [String] -> [(Int, Int)]

Função que recebe a parte das coordenadas e as transforma para o formato (x,y).

 >>> leCoordenadas ["1 1","3 4"]
[(1,1),(3,4)]

Funções Principais

Funções que inserem o boneco e as caixas no mapa

altera :: [String] -> [(Int, Int)] -> [String]

Insere as caixas e o boneco nas posições respetivas (de acordo com as coordenadas dadas).

>>> altera ["##########","##     ###","## ###   #","#        #","# ..#   ##","##..#   ##","##########"] [(2,3),(2,4),(4,3),(6,2),(7,3)]
["##########","##     ###","##H###   #","# o H  H #","# ..# H ##","##..#   ##","##########"]

inserir :: String -> [String] -> Int -> [String]

Muda uma linha do tabuleiro por uma nova, no formato correto.

>>> inserir "##o##" ["#####","## ##","#####"] 1
["#####","##o##","#####"]

alteraboneco :: (Int, Int) -> [String] -> String

Esta função recebe as coordenadas do boneco e devolve a linha na qual ele se encontra, já com o caractere 'o' no sítio correto.

>>> alteraboneco (2,1) ["#####","## ##","#####"] 
"##o##"

alteracaixa :: [String] -> [(Int, Int)] -> [String]

Recebe um tabuleiro e uma lista de coordenadas e insere um 'H' ou um 'I' nas posições das caixas.

>>> alteracaixa ["#####","## ##","#####"] [(2,1)] 
["#####","##H##","#####"]

Funções que removem '#' redundantes

remove :: (Int, Int) -> [String] -> [String] -> [String]

Retira os caracteres '#' que forem redundantes.

>>>  remove (0,0) ["##########","##     ###","## ###   #","#        #","# ..#   ##","##..#   ##","##########"] ["##########","##     ###","## ###   #","#        #","# ..#   ##","##..#   ##","##########"]
[" #######  "," #     ###","## ###   #","#        #","# ..#   ##","##..#   # "," ######## "]

remove2 :: Char -> (Int, Int) -> [String] -> String

Verifica se a remoção do caractere '#' é possível. Caso seja possível, remove-o; caso contrário, mantém-no.

>>>  remove2 '#' (1,1) ["#####","## ##","#####"] 
"#"

>>>  remove2 '#' (0,0) ["#####","## ##","#####"] 
" "