Main

Description

O objetivo deste módulo é criar o mapa do nosso projeto Bomberman. A função mapa receberá uma dimensão e uma seed (cujo objetivo

será gerar uma lista de números aleatórios que corresponde a tijolos, possivelmente a cobrir power ups, ou espaços vazios. O output

desta função é uma lista de Strings onde está contida toda a informação do jogo em relação ao espaço onde os jogadores estarão .

Documentation

main :: IO ()

Função que recebe os inputs do terminal e utiliza a função mapa para imprimir no stdout a lista de Strings resultantes.

mapa :: Int -> Int -> [String]

A função mapa cria a lista de strings. Optámos por dvidir este processo em duas partes, a construção do mapa propriamente dito, e a

concatenação dessa lista de Strings com a informação relativa aos powerups. Ambas essas subdivisões da função mapa fazem uso da lista

de números aleatórios gerados a partir da seed com um número de elementos que depende da dimensão d.

>mapM_ putStrLn (mapa 9 0)

#########

? #

 ?#

? #

#########
+ 5 2
+ 3 3
! 5 5

emptyMap :: Int -> Int -> String

Função que gera o mapa do jogo vazio, em forma de String. Os espaços vazios fixos foram marcados com um X de modo a serem

facilmente substituídos por outros caracteres. Optámos por manter, nesta fase, o mapa em formato String porque isso facilita a nossa

forma de trabalhar com a informação, tendo em conta que nós pretendemos utilizar funções que percorram os elementos do mapa.

Como o caso dos mapas de dimensão 5 é especial, em que há sobreposição de espaços vazios fixos, criámos uma definição própria para ele.

Exemplo:

> emptyMap 9 9

"######### X#  # X##########"

Apesar deste ser o output, podemos tentar visualizar este resultado da seguinte forma:

"#########
 
  X#
 
   #
 
  X#
 
 #########"

 

rewriteL :: String -> [Int] -> String

Função que traduz os X criados na função emptyMap em espaços vazios, e substitui os espaços que não estavam marcados por caracteres

resultantes da lista de números aleatórios. Note-se que a definição rewriteL ls [] = rewriteL ls [1] existe para prevenir que a função

receba uma lista de números vazia, apesar de sabermos que, quando esta fica sem elementos, não será necessaŕio mais nenhum deles, e por isso

é irrelevante o elemento que adicionamos à lista para a função poder lidar com ela.

> rewriteL (emptyMap 9 9) (take 28 $ randomRs (0,99) (mkStdGen s))

"#########? # ?#? ##########"

translate :: Int -> Char

Função que traduz os numeros aleatórios em símbolos para o mapa.

>translate 10

?

lilstring :: String -> Int -> [String]

Função que agrupa os elementos de uma string st em strings de comprimento n (que será igual ao comprimento do lado do mapa), pela mesma

ordem em que aparecem em st. Sendo aplicada esta função, a informação toma a forma do output final, [String].

powers :: String -> [Int] -> Int -> [String]

O papel desempenhado pela função powers é o de adicionar por baixo do mapa a informação relativa aos power ups. Como os bombs têm de vir primeiro

do que os flames, decidimos criar duas funções semelhantes que constroem as Strings relativas a essa informação e cujos outputs são concatenados na powers

>powers (rewriteL (emptyMap 9 9) ls) ls 9

["+ 5 2","+ 3 3","! 5 5"]

nota: seja ls = take ((d-2)^2 - (div (d-3) 2)^2 - 12) $ randomRs (0,99) (mkStdGen 0)

powers1 :: String -> [Int] -> Int -> Int -> [String]

Esta função faz uso de um contador n, a partir do qual podemos calcular as coordenadas, em função de d, dos power ups que encontramos.

powers2 :: String -> [Int] -> Int -> Int -> [String]

Função semelhante à powers1, com a única diferença de gerar as strings respeitantes aos power ups flames.

dozeVazios :: Int -> [Int]

Lista das 12 células, em função de d, que estarão sempre vazias.