Main

Description

O objetivo desta tarefa é implementar criar uma função que comprima o estado do jogo em uma String e outra que reverta o processo, ou seja, que , a partir da String

resultante da codificação, volte a construir a exata lista de String correspondente ao estado em que o jogo se encontrava. Quanto menor for a String resultante da

compressão, melhor, dado que se poupa mais informação.

Documentation

encode :: [String] -> String

A função encode divide a tarefa de comprimir o estado do mapa em duas partes. A primeira é de comprimir o mapa propriamente dito, a segunda é decomprimir a

informação das Strings abaixo do mapa, relativa a power ups, jogadores e bombas.

Outro pormenor da nossa função encode é que guarda no início da String a dimensão do mapa, a fim de ser mais fácil de trabalhar na descodificação desta String.

decode :: String -> [String]

Como seria de esperar, a função decode é fortemente influenciada pelo raciocínio e estratégia adotados na encode, dado que tem de, de certa forma, realizar

o processo inverso. Assim, esta função também divide o seu papel em duas partes, a de construir o mapa e a de acrescentar por baixo a informação relativa a

power ups, bombas e jogadores.

main :: IO ()

A função main recebe como input um estado do jogo (geralmente importado dos exemplos na pasta de testes através do stdin) e um comando, -e para encode ou

d para decode, que determinarão a função a utilizar na informação importada.

inserirPowerups :: [String] -> [String] -> [String]

O primeiro passo da compressão é a função inserirPowerups que insere no mapa os powerups (com caracteres diferentes caso estejam por baixo de tijolos ou

não). Assim, não há a necessidade de ter as Strings que se encontram depois do mapa, à exceção de alguma informação relativa a bombas e as Strings dos

jogadores, que não inserimos no mapa por ser incómodo inserir dígitos no mapa, uma vez que utilizamos dígitos no nosso método de compressão.

tiraLinhas :: [String] -> String

A função tiralinhas rempve as Strings com informação relativa a powerups do mapa, e retira a informação relativa à posição das bombas.

stringInfo :: String -> String

A função stringinfo recolhe a informação que resta das bombas (em strings que deixámos com o Char > início) e a informação dos jogadores (em linhas

separadas por barras (/) ).

stringMapa :: String -> String

A função stringMapa, a partir de uma string com a informação do mapa e a restande das bombas e jogadores, devolve apenas a parte relaiva ao mapa, que é

processada por uma função diferente do resto.

removeCard :: String -> String

Esta função retira os cardinais da string relativa ao mapa, uma vez que estes estão sempre nos mesmos sítios e os podemos recriar nas mesmas posições.

removeVazios :: String -> Int -> String

Esta função substitui os chars ' ' por números que representam a quantidade de espaços seguidos que se encontram nessa posição da string.

encodeBricks :: String -> String -> String

Esta função procede da mesma forma que a removeVazios, mas para os tijolos (caracter ?) e substitui os aglomerados destes mesmos caracteres por letras

(e.g. a=1, b=2, c=3 ...)

encodePadrao :: String -> String

A função encodePadrão, dada a string resultante da compressão já feita, reconhece padrões de dois caracteres. A esses padrões faz corresponder uma letra

maiúscula.

encodePadrao2 :: String -> String

A função encodePadrao2 procede das mesma forma que a encodePadrao, no entanto reconhece conjuntos de duas letras máiúsculas e substitui-os por um símbolo.

encodeInfo :: String -> String

A função encodeInfo efetua um processo semelhante ao da encodePadrao mas destinado e adaptado à informação relativa a bombas e jogadores que não foi

incluida no mapa.

decodeInfo :: String -> String

A função decodeInfo faz a substituição inversa da encodeInfo.

decodePadrao2 :: String -> String

A função decodePadrao2 efetua as substituições inversas da encodePadrao2

decodePadrao :: String -> String

A função decodePadrao efetua as substituições inversas da encodePadrao.

decodeBricks :: String -> String

A função decodeBricks substitui os as letras que encontra pela quantidade de tijolos a que corresponde essa letra.

addVazios :: String -> String

Esta função substitui os dígitos na string pela quantidade correspondente de espaços seguidos.

gerarMapaCard :: Int -> Int -> String

Esta função, a partir da dimensão do mapa (armazenada no inicio da string resultante da decode), gera um mapa vazio (apenas com os cardinais).

fillMapaCard :: String -> String -> String

A função fllMapaCard procede à substituição dos espaços do mapa vazio em que não há cardinais pelos caracteres da string a ser descodificada, cujo número de

elementos coincide com o número de caracteres que serão precisos, naturalmente.

lilstring :: String -> Int -> [String]

A função lilstring, a partir de uma dimensão d, divide uma String numa lista de d strings, com d elementos cada uma.

constroiMapa :: [String] -> [String]

A função constroiMapa gera o output final da decode, retirando os power ups e as bombas de dentro do mapa e registando a informação relativa

aos mesmos em strings que o seguem, como num normal estado do mapa.

unlineInfo :: String -> String -> [String]

A função unlineInfo, a partir da string codificada separa a informação relativa a cada bomba e a cada jogador, gerando uma lista de Strings que

serve para ser junta com a informação retirada do mapa.

gerarInfoMapa1 :: [String] -> Int -> [String]

Esta função tem como objetivo gerar uma lista de Strings com toda a informação relativa aos power ups bombs, recorrendo a uma função auxiliar, que

concretiza este processo para cada linha.

powersNumaString1 :: String -> Int -> Int -> [String]

Esta função serve de auxiliar à gerarInfoMapa1, gerando também uma lista de Strings, dado que podem haver vários powerups numa linha.

gerarInfoMapa2 :: [String] -> Int -> [String]

Esta função é exatamente igual à gerarInfoMapa1, no entanto trata dos power ups flames. Procedemos desta forma, porque os powerups têm de surgir

organizados no mapa (primeiro bombs, depois flames)

powersNumaString2 :: String -> Int -> Int -> [String]

Função que serve de auziliar à gerarInfoMapa2 .

gerarInfoMapa3 :: [String] -> Int -> [String]

Outra função semelhante à gerarInfoMapa1, que trata da informação relativa a bombas.

powersNumaString3 :: String -> Int -> Int -> [String]

Esta função gera informação relativa às bombas que se encontrem numa dad string do mapa.

unirBombas :: [String] -> Int -> [String]

A função unir bombas aproveita a informação recuperada das bombas colocadas no mapa e volta a unir essas Strings com as relativas à informação

alheia à localização das bombas, que foram deiixadas de fora.

procurarBomba :: String -> [String] -> Int -> Int -> String

A função procurarBomba percorre o mapa em busca da string com a informação restante de cada bomba que é retirada do mapa.

retirarPowerups :: [String] -> [String]

A função retirarPowerups percorre o mapa e chama uma função que retira os caracteres dos powerups de cada String.

subsPowerups :: String -> String

Esta função percorre uma string e substitui qualquer caracter relativo a power ups pelo caracter que lá se encontrava inicialmente, o que é

possível porque definimos caracteres diferentes para power ups debaixo de tijolos e power ups destapados.

pOrden :: String -> Int

Esta função converte a String relativa à ordenada de um objeto num Int, para ser mais cómodo trabalhar com esta informação.

pAbcis :: String -> Int

Esta função converte a String relativa à abcissa de um objeto num Int, para ser mais cómodo trabalhar com esta informação.

pAbcisSt :: String -> String

Esta função, a partir de uma String relativa a um dado objeto (jogador, power up, bomba..), a parte que corresponde à abcissa da posição desse objeto

no mapa.

pOrdenSt :: String -> String

Esta função, a partir de uma String relativa a um dado objeto (jogador, power up, bomba..), a parte que corresponde à ordenada da posição desse objeto

no mapa.

howMuch :: Char -> String -> Int

Esta função faz corresponder a uma letra do alfabeto (minúscula) o número da sua ordem no alfabeto (e.g. a=1, b=2, c=3...).