Main

Description

Neste módulo procura-se que o jogo em criação passe a ser capaz de alterar o seu estado com base num acontecimento. O acontecimento nesta tarefa será especificamente a passagem do tempo onde o jogo terá de se adaptar com base nos eventos que ocorrem no mapa: Exemplos disto serão os timers das bombas, que com a passagem do tempo serão reduzidos assim como a alteração física do mapa a partir do instante (d-2)² onde d representa a dimensão do mesmo e instante onde é invocada uma espiral que vai fechando o mapa.

Documentation

type Mapa

Arguments

= [String]
Representação do estad do jogo em lista de strings.

ex14 ######### # # # ######### * 7 6 1 1 1 * 7 7 0 1 10 0 7 5 1 7 5

type Linha

Arguments

= String
Uma linha da representação doe stado do jogo no tipo Mapa

type PowerType

Arguments

= Char
Char distintivo de um tipo de powerup.

type Range

Arguments

= Int
Raio de uma bomba.

type Abcissa

Arguments

= Int
Abcissa de um espaço do mapa.

type Ordenada

Arguments

= Int
Ordenada de um espaço do mapa.

type Tempo

Arguments

= Int
Instantes de tempo que restam para o jogo acabar.

type Cords

Arguments

= (Abcissa, Ordenada)
Coordenadas de um espaço no mapa.

type Dim

Arguments

= Int
Dimensão do mapa.

type Jogador

Arguments

= Int
Número que representa um dado jogador.

ln :: Foldable t => t String -> IO ()

avanca :: Mapa -> Tempo -> Mapa

mapa1 :: [[Char]]

main :: IO ()

reduce :: Mapa -> Mapa

Exemplo: reduce ex14 ################ ################ * 7 6 1 1 1 * 7 6 1 1 0 * 7 7 0 1 10 * 7 7 0 1 9 0 7 5 0 7 5 1 7 5 1 7 5

Função que percorre o mapa e, a cada bomba, reduz o tempo do contador em 1.

tickBomb :: Linha -> Linha

Função que reduz o contador na string da bomba.

explosionsMap :: Mapa -> Mapa -> [Mapa]

Função que percorre o mapa para encontrar bombas que têm de explodir, levando outro argumento com o mapa, que vai sendo alterado à medida que vão havendo explosões.

mergeMaps :: [Mapa] -> Mapa -> Mapa

Função que funde mapas resultantes de explosões que ocorreram em simultâneo.

howMany :: PowerType -> Linha -> Int

explode :: Mapa -> Linha -> Mapa

Função que, a partir do estado do mapa e de uma string de uma bomba prestes a explodir, gera o mapa depois da explosão.

reduce ex14 explode ex14 ####################### # -> # # ####################### * 7 6 1 1 1 * 7 6 1 1 0 * 7 7 0 1 9 * 7 7 0 1 10 * 7 7 0 1 9 0 7 5 0 7 5 0 7 5 1 7 5 1 7 5 1 7 5

flameUp :: Mapa -> Abcissa -> Ordenada -> Range -> Mapa

Função que explode os espaços do mapa acima da bomba e chama as funções para efetuar as alterações necessárias no estado do jogo.

flameDown :: Mapa -> Abcissa -> Ordenada -> Range -> Mapa

Função que explode os espaços do mapa abaixo da bomba e chama as funções para efetuar as alterações necessárias no estado do jogo.

flameRight :: Mapa -> Abcissa -> Ordenada -> Range -> Mapa

Função que explode os espaços do mapa à direita da bomba e chama as funções para efetuar as alterações necessárias no etsado do jogo.

flameLeft :: Mapa -> Abcissa -> Ordenada -> Range -> Mapa

Função que explode os espaços do mapa à esquerda da bomba e chama as funções para efetuar as alterações necessárias no etsado do jogo.

espiral :: Mapa -> Tempo -> Mapa

Função que coloca as pedras e esmaga o que estiver nessas coordenadas.

cordLeft :: Cords -> Abcissa -> [Cords]

Função auxiliar que cria a lista de coordenadas da espiral da direita para a esquerda.

cordRight :: Cords -> Abcissa -> [Cords]

Função com a mesma funcionalidade que a cordLeft mas com diferente direção.

cordUp :: Cords -> Ordenada -> [Cords]

Atua como as outras cord mas preenche a lista de baixo para cima.

cordDown :: Cords -> Ordenada -> [Cords]

Preenche a lista de cima para baixo.

quadrado :: Cords -> Int -> [Cords]

Função que engloba todas as funções cord e que gera ordenadamente a lista de coordenadas que serão destruidas pela espiral.

listaSpiral :: Cords -> Dim -> [Cords]

Função que gera uma lista de pares de coordenadas que respresentam a ordem pela qual o mapa será preenchido por pedras.

placeRock :: Mapa -> Tempo -> Mapa

Função que coloca a pedra relativamente a um instante do jogo.

eraseObj :: Mapa -> Tempo -> Mapa

Função que "esmaga" o que estiver numas dadas coordenadas (onde será colocad auma pedra)

removePower :: Mapa -> Abcissa -> Ordenada -> Mapa

Função que remove o powerup numas dadas coordenadas.

removePlayer :: Mapa -> Abcissa -> Ordenada -> Mapa

Função que remove o jogador numas dadas coordenadas.

blowTimer :: Mapa -> Abcissa -> Ordenada -> Mapa

Função que coloca o timer da bomba numas coordenadas a 1.

stringTimer :: Linha -> Linha

Função que coloca o timer da string de uma bomba a 1 (a menos que esteja para explodir).

removeBomb :: Mapa -> Abcissa -> Ordenada -> Mapa

Função que remove a bomba de umas dadas coordenadas (retira-a do Mapa, [String]).

isTherePower :: Mapa -> Abcissa -> Ordenada -> Bool

Função que testa se há um powerup num dado espaço do mapa.

isTherePlayer :: Mapa -> Abcissa -> Ordenada -> Bool

Função que testa se há um jogador num dado espaço do mapa.

isThereBomb :: Mapa -> Abcissa -> Ordenada -> Bool

Função que testa se há um powerup num dado espaço do mapa.

pOrden :: Linha -> Ordenada

Esta pequena função traduz a função pOrdenSt num Int, de forma a podermos trabalhar com esta informação em índices, por exemplo.

pAbcis :: Linha -> Abcissa

O papel desempenhado pela pAbcis é análogo ao da pOrden, mas diz respeito à abcissa de um dado objeto (e.g. jogador, bomba, power up..)

pAbcisSt :: Linha -> String

Esta função obtem a parte da string correspondente a um jogador, ou outro objeto com coordenadas, em que se encontra a informação sobre a abcissa

das coordenadas desse mesmo objeto. Como esta informação começa sempre no 3º caracter de cada string, apenas temos de considerar o quão grande é o

número relativo a esta mesma coordenada.

pOrdenSt :: Linha -> String

Apesar de esta função ter exatamente o mesmo papel que a pAbcisSt, a ordenada dos objetos vem em segundo lugar, e, por isso, os caracteres da string

que correspondem a esta informação também dependem da quantidade de caracteres utilizados para descrever a abcissa.

range :: Linha -> Int

Função que retorna o range da bomba descrita numa dada string.

cordsT :: Linha -> (Int, Int)