Main

Description

O objetivo desta tarefa é a implementação propriamente funcional do jogo. Embora nas outras tarefas já tenham sido desenvolvidas soluções par aa reação a comandos e passagem do tempo, continua a ser necessária a utilização de uma plataforma de desenvolvimento de uma interface gráfica, reproduzida a partir de cada estado do jogo. Compilada esta tarefa, deve resultar o produto final do projeto.

Documentation

type Brick = (Int, Int)

(abcissa, ordenada)

type Player = (Int, Int, Int, Int, Int)

(numero,abcissa,ordenada,bombs,flames)

type Powerup = (Char, Int, Int)

(powerup, abcissa, ordenada)

type Bomb = (Int, Int, Int, Int, Int)

(abcissa, ordenada, player, range, timer)

type Estado = (Mapa, [Brick], [Brick], [Powerup], [Bomb], [Player], Float, Screen)

(lista de strings, pedras, tijolos, powerups, bombas, players, tempo, informação sobre fase do programa)

type Seed = Int

Seed geradora do mapa.

type Tempo = Int

Instantes que faltam para o jogo acabar

type Jogador = Int

Número representativo de um jogador.

type Screen = String

String representativa do ecrã em que o programa do jogo se encontra.

estadoInicial :: Seed -> Estado

O estado inicial do jogo.

estadoInicial2 :: Dim -> Seed -> Int -> Estado

O estado inicial do jogo, depois de escolhidos a dimensão e o numero de players.

fr :: Int

Frame rate

dm :: Display

Display mode

main :: IO ()

Função principal que invoca o jogo.

reageEvento :: Event -> Estado -> Estado

Função que altera o estado do jogo quando acontece um evento. ---

reageDim :: Event -> Estado -> Estado

Função que faz mover o círculo mediante o input do jogador no men de seleção da dimensão.

enterDim :: Estado -> Estado

Função que seleciona a dimensão pretendida pelo jogador.

reagePls :: Event -> Estado -> Estado

Função que faz mover o círculo mediante o input do jogador no menu de seleção do numero de players.

enterPls :: Estado -> Estado

Função que seleciona o numero de jogadores pretendido pelo jogador.

reageEvento1 :: Event -> Estado -> Estado

Função que controla os movimentos dos jogadores.

reageOver :: Event -> Estado -> Estado

Função que faz com que, premida alguma tecla após o jogo acabar, o jogo volte ao início.

newseed :: Seed -> Seed

Função que gera uma nova seed a partir da anterior.

reageTempo :: Float -> Estado -> Estado

Função que altera o estado do jogo quando o tempo avança n segundos.

botTempo :: Mapa -> Tempo -> Mapa

Função que recebe o estado e devolve-o com as alterações de acordo com o input dos bots.

desenhaEstado :: [Picture] -> [Picture] -> [Picture] -> [Picture] -> Estado -> Picture

Função que desenha o jogo, levando como argumento as listas de imagens relativas a diferentes componentes do jogo.

pedras :: Picture -> Dim -> Estado -> Picture

Esta função gera uma imagem com todas as pedras contidas no mapa.

tijolos :: Picture -> Dim -> Estado -> Picture

Esta função gera uma imagem com todos os tijolos contidos no mapa.

powersImg :: Picture -> Picture -> Dim -> [Powerup] -> Estado -> [Picture]

Esta função gera uma imagem com todos os powerups contidos no mapa.

bombasImg :: Picture -> Dim -> Estado -> Picture

Esta função gera uma imagem com todas as bombas contidas no mapa.

playersImg :: [Picture] -> Dim -> Estado -> Picture

Esta função gera uma imagem com todos os jogadores contidos no mapa.

scoreB :: [Picture] -> Estado -> Picture

sbPlayers :: [Picture] -> [Player] -> [Picture]

burn :: [Picture] -> Estado -> [Picture]

Função que coloca as imagens das chamas.

burn2 :: [Picture] -> Int -> [Flames] -> [Picture]

Função que coloca a imagem de cada espaço a arder.

drawDim :: [Picture] -> Estado -> Picture

Função que desenha o menu de seleção da dimensão do mapa.

redCircleD :: Screen -> Picture

Função que desenha o círculo que ajuda a selecionar uma opção.

drawPls :: [Picture] -> Estado -> Picture

Função que desenha o menu de seleção do nú mero de jogadores.

redCircleP :: Screen -> Picture

Função que desenha o círculo que ajuda a selecionar uma opção.

gameover :: Estado -> [Picture] -> [Picture] -> Picture

Função que desenha na janela o vencedor no fim do jogo.

winnerPlayer :: [Player] -> [Picture] -> [Picture]

Função que deteta o jogador que ganhou o jogo e imprime a sua imagem.

resized :: Picture -> Dim -> Picture

Função que redimensiona uma imagem importada de acordo om uma dimensão do mapa.

r :: Dim -> Float

Função que, a partir da dimensão do mapa, indica o rácio necessário para redimensionar as imagens importadas (que têm sempre 85px, por escolha nossa)

squareSide :: Dim -> Float

Dimensão do lado de cada quadrado do mapa, em função da dimensão do mapa de jogo.

placeMe :: Dim -> Abcissa -> Ordenada -> Picture -> Picture

Função que a partir de umas coordenadas e de uma dimensão demapa coloca corretamente imagens na janela.

redCircle1 :: Picture

Cículo utilizado nos menus de seleção.

tempo :: Estado -> Picture

Função que mostra o visor como numero de instantes até o jogo terminar.

inserePlayers :: Mapa -> Int -> Dim -> [String]

Função que insere num mapa gerado pela tarefa1 um número especificado de jogadores.

geraEstado :: Mapa -> Float -> Screen -> Estado

Função que, a partir de uma lista de strings representativa de um mapa e do tempo, gera uma estrutura do tipo Estado, anteriormente definido.

retiraRocks :: Ordenada -> Mapa -> [Brick]

Função que retira do mapa as pedras em forma de túpulo.

retiraRocksRow :: Ordenada -> Abcissa -> Linha -> [Brick]

Função que retira de cada linha do mapa as pedras, em forma de túpulo.

retiraBricks :: Ordenada -> Mapa -> [Brick]

Função que retira do mapa os tijolos em forma de túpulo.

retiraBricksRow :: Ordenada -> Abcissa -> Linha -> [Brick]

Função que retira de cada linha do mapa os tijolos, em forma de túpulo.

retiraBombs :: Mapa -> [Bomb]

Função que retira do mapa as bombas colocadas.

retiraPups :: Mapa -> [Powerup]

Função que a partir do estado do jogo obtém informação dos powerups.

tuplePower :: Linha -> Powerup

Função que converte num túpulo uma string com informação de dum dado powerup.

geraPlayers :: Mapa -> [Player]

Função que gera a partir de um estado do jogo (no formato da tarefa 2) uma lista de túpulos Player.

tuplePlayer :: Linha -> Player

Função que, a partir de uma string de um jogador, gera o túpulo relativo ao mesmo.

takeLString :: Estado -> Mapa

Função que retira de um estado a lista de strings associada à representação do mapa.

takeRocks :: Estado -> [Brick]

Função que retira de um estado a lista de pedras no mapa.

takeBricks :: Estado -> [Brick]

Função que retira de um estado a lista de tijolos no mapa.

takePowers :: Estado -> [Powerup]

Função que retira de um estado a lista de powerups no mapa, em forma de túpulos.

takeBombs1 :: Estado -> [Bomb]

Função que retira de um estado a lista de bombas no mapa, em forma de túpulos.

takePlayers1 :: Estado -> [Player]

Função que retira de um estado a lista de jogadores no mapa, em forma de túpulos.

takeTime :: Estado -> Float

Função que retira de um estado o tempo do jogo.

takeMen :: Estado -> String

Função que retira de um estado o elemento que especifica o tipo de interface a representar.

takeChar :: Powerup -> Char

Função que descobre o tipo de powerup representado pelo túpulo (obtém o seu Char representativo)

takeX :: Powerup -> Abcissa

Função que obtém a abcissa de um determinado powerup.

takeY :: Powerup -> Ordenada

Função que obtém a ordenada de um determinado powerup.

howMany1 :: Char -> Linha -> Int

Função que conta o número de vezes que um certo caracter ocorre numa string, com utilidade, por exemplo, no caso de saber quantos powerups de determinado tipo tem um jogador.

isInt :: Float -> Bool

Função que verifica se um float é inteiro.

deci :: Float -> Float

Função que devolve a parte decimal de um float